refactor(agents): decompose mod.rs into dedicated submodules
Decompose the monolithic `src/agents/mod.rs` into single-responsibility
submodules while preserving 100% backward compatibility:

- Move file I/O, JSON/JSONC/TOML, and path helpers to `src/agents/fs.rs`
- Move rule generation and instruction file maintenance to `src/agents/rules.rs`
- Move global Git hook installation and logic to `src/agents/hooks.rs`
- Move agent registry, lookup, and migration to `src/agents/registry.rs`
- Move core traits and context structs to `src/agents/traits.rs`
- Relocate agent implementations under `src/agents/integrations/`

All public API paths remain re-exported from `src/agents/mod.rs`.
Zero functional or behavioral logic changes.
