+++ src/agents/integrations/antigravity.rs
@@ -17,10 +17,7 @@ use serde_json::json;
 
 use crate::errors::Result;
 
-use super::{
-    backup_config_file, load_json_file, load_json_file_strict, safe_write_json_file,
-    AgentIntegration, DoctorCounters, HealthcheckContext, InstallContext,
-};
+use super::*;
 
 /// Google Antigravity agent.
 pub struct AntigravityIntegration;
