+++ src/agents/integrations/augment.rs
@@ -21,10 +21,7 @@ use serde_json::json;
 
 use crate::errors::Result;
 
-use super::{
-    backup_and_write_json, backup_config_file, load_json_file, load_json_file_strict,
-    safe_write_json_file, AgentIntegration, DoctorCounters, HealthcheckContext, InstallContext,
-};
+use super::*;
 
 /// `AugmentCode` (`auggie`) agent.
 pub struct AugmentIntegration;
