+++ src/agents/integrations/cline.rs
@@ -1,6 +1,6 @@
-//! Roo Code agent integration.
+//! Cline agent integration.
 //!
-//! Handles registration of the tokensave MCP server in Roo Code's
+//! Handles registration of the tokensave MCP server in Cline's
 //! `cline_mcp_settings.json` under the `mcpServers.tokensave` key.
 
 use std::path::{Path, PathBuf};
@@ -9,44 +9,27 @@ use serde_json::json;
 
 use crate::errors::Result;
 
-use super::{
-    backup_and_write_json, backup_config_file, load_json_file, load_json_file_strict,
-    safe_write_json_file, AgentIntegration, DoctorCounters, HealthcheckContext, InstallContext,
-    InstallScope,
-};
+use super::*;
 
-/// Roo Code agent.
-pub struct RooCodeIntegration;
+/// Cline agent.
+pub struct ClineIntegration;
 
-/// Returns the Roo Code VS Code extension global storage directory.
-fn roo_ext_dir(home: &Path) -> PathBuf {
-    super::vscode_data_dir(home).join("User/globalStorage/rooveterinaryinc.roo-cline")
+/// Returns the Cline VS Code extension global storage directory.
+fn cline_ext_dir(home: &Path) -> PathBuf {
+    super::vscode_data_dir(home).join("User/globalStorage/saoudrizwan.claude-dev")
 }
 
-/// Roo Code MCP settings path for this install: the extension global storage
-/// for a global install, `<project>/.roo/mcp.json` for `--local`.
-fn roo_settings_path(ctx: &InstallContext) -> PathBuf {
-    match &ctx.scope {
-        InstallScope::Global => roo_ext_dir(&ctx.home).join("settings/cline_mcp_settings.json"),
-        InstallScope::Local { project_path } => project_path.join(".roo/mcp.json"),
-    }
-}
-
-impl AgentIntegration for RooCodeIntegration {
+impl AgentIntegration for ClineIntegration {
     fn name(&self) -> &'static str {
-        "Roo Code"
+        "Cline"
     }
 
     fn id(&self) -> &'static str {
-        "roo-code"
-    }
-
-    fn supports_local(&self) -> bool {
-        true
+        "cline"
     }
 
     fn install(&self, ctx: &InstallContext) -> Result<()> {
-        let settings_path = roo_settings_path(ctx);
+        let settings_path = cline_ext_dir(&ctx.home).join("settings/cline_mcp_settings.json");
 
         if let Some(parent) = settings_path.parent() {
             std::fs::create_dir_all(parent).ok();
@@ -81,35 +64,35 @@ impl AgentIntegration for RooCodeIntegration {
         crate::agent_note!();
         crate::agent_note!("Setup complete. Next steps:");
         crate::agent_note!("  1. cd into your project and run: tokensave init");
-        crate::agent_note!("  2. Restart VS Code — tokensave tools are now available in Roo Code");
+        crate::agent_note!("  2. Restart VS Code — tokensave tools are now available in Cline");
         Ok(())
     }
 
     fn uninstall(&self, ctx: &InstallContext) -> Result<()> {
-        let settings_path = roo_settings_path(ctx);
+        let settings_path = cline_ext_dir(&ctx.home).join("settings/cline_mcp_settings.json");
         uninstall_mcp_server(&settings_path);
 
         crate::agent_note!();
-        crate::agent_note!("Uninstall complete. Tokensave has been removed from Roo Code.");
+        crate::agent_note!("Uninstall complete. Tokensave has been removed from Cline.");
         crate::agent_note!("Restart VS Code for changes to take effect.");
         Ok(())
     }
 
     fn healthcheck(&self, dc: &mut DoctorCounters, ctx: &HealthcheckContext) {
-        crate::agent_note!("\n\x1b[1mRoo Code integration\x1b[0m");
+        crate::agent_note!("\n\x1b[1mCline integration\x1b[0m");
         doctor_check_settings(dc, &ctx.home);
     }
 
     fn is_detected(&self, home: &Path) -> bool {
-        roo_ext_dir(home).is_dir()
+        cline_ext_dir(home).is_dir()
     }
 
     fn primary_config_path(&self, home: &Path) -> Option<PathBuf> {
-        Some(roo_ext_dir(home).join("settings/cline_mcp_settings.json"))
+        Some(cline_ext_dir(home).join("settings/cline_mcp_settings.json"))
     }
 
     fn has_tokensave(&self, home: &Path) -> bool {
-        let settings_path = roo_ext_dir(home).join("settings/cline_mcp_settings.json");
+        let settings_path = cline_ext_dir(home).join("settings/cline_mcp_settings.json");
         if !settings_path.exists() {
             return false;
         }
@@ -124,7 +107,7 @@ impl AgentIntegration for RooCodeIntegration {
 // Uninstall helpers
 // ---------------------------------------------------------------------------
 
-/// Remove MCP server entry from Roo Code's `cline_mcp_settings.json`.
+/// Remove MCP server entry from Cline's `cline_mcp_settings.json`.
 fn uninstall_mcp_server(settings_path: &Path) {
     if !settings_path.exists() {
         crate::agent_note!("  {} not found, skipping", settings_path.display());
@@ -180,13 +163,13 @@ fn uninstall_mcp_server(settings_path: &Path) {
 // Healthcheck helpers
 // ---------------------------------------------------------------------------
 
-/// Check Roo Code's `cline_mcp_settings.json` has tokensave MCP server registered.
+/// Check Cline's `cline_mcp_settings.json` has tokensave MCP server registered.
 fn doctor_check_settings(dc: &mut DoctorCounters, home: &Path) {
-    let settings_path = roo_ext_dir(home).join("settings/cline_mcp_settings.json");
+    let settings_path = cline_ext_dir(home).join("settings/cline_mcp_settings.json");
 
     if !settings_path.exists() {
         dc.warn(&format!(
-            "{} not found — run `tokensave install --agent roo-code` if you use Roo Code",
+            "{} not found — run `tokensave install --agent cline` if you use Cline",
             settings_path.display()
         ));
         return;
@@ -202,7 +185,7 @@ fn doctor_check_settings(dc: &mut DoctorCounters, home: &Path) {
         ));
     } else {
         dc.fail(&format!(
-            "MCP server NOT registered in {} — run `tokensave install --agent roo-code`",
+            "MCP server NOT registered in {} — run `tokensave install --agent cline`",
             settings_path.display()
         ));
     }
