+++ src/agents/integrations/kimi.rs
@@ -14,10 +14,7 @@ use serde_json::json;
 
 use crate::errors::{Result, TokenSaveError};
 
-use super::{
-    backup_and_write_json, backup_config_file, load_json_file, load_json_file_strict,
-    safe_write_json_file, AgentIntegration, DoctorCounters, HealthcheckContext, InstallContext,
-};
+use super::*;
 
 /// Moonshot Kimi CLI agent.
 pub struct KimiIntegration;
