+++ src/agents/integrations/kiro.rs
@@ -18,11 +18,7 @@ use serde_json::json;
 
 use crate::errors::{Result, TokenSaveError};
 
-use super::{
-    backup_and_write_json, backup_config_file, load_json_file, load_json_file_strict,
-    safe_write_json_file, AgentIntegration, DoctorCounters, HealthcheckContext, InstallContext,
-    InstallScope,
-};
+use super::*;
 
 /// Kiro agent.
 pub struct KiroIntegration;
