+++ src/agents/integrations/mod.rs
@@ -1,0 +1,46 @@
+pub mod antigravity;
+pub mod augment;
+pub mod claude;
+pub mod cline;
+pub mod codex;
+pub mod copilot;
+pub mod cursor;
+pub mod droid;
+pub mod gemini;
+pub mod grok;
+pub mod kilo;
+pub mod kimi;
+pub mod kiro;
+pub mod opencode;
+pub mod pi;
+pub mod plank;
+pub mod qwen;
+pub mod roo_code;
+pub mod vibe;
+pub mod zed;
+
+pub use antigravity::AntigravityIntegration;
+pub use augment::AugmentIntegration;
+pub use claude::ClaudeIntegration;
+pub use cline::ClineIntegration;
+pub use codex::CodexIntegration;
+pub use copilot::CopilotIntegration;
+pub use cursor::CursorIntegration;
+pub use droid::DroidIntegration;
+pub use gemini::GeminiIntegration;
+pub use grok::GrokIntegration;
+pub use kilo::KiloIntegration;
+pub use kimi::KimiIntegration;
+pub use kiro::KiroIntegration;
+pub use opencode::OpenCodeIntegration;
+pub use pi::PiIntegration;
+pub use plank::PlankIntegration;
+pub use qwen::QwenIntegration;
+pub use roo_code::RooCodeIntegration;
+pub use vibe::VibeIntegration;
+pub use zed::ZedIntegration;
+
+use crate::agents::traits::*;
+use crate::agents::fs::*;
+use crate::agents::rules::*;
+use crate::agents::*;
