+++ src/agents/integrations/opencode.rs
@@ -12,13 +12,7 @@ use serde_json::json;
 
 use crate::errors::Result;
 
-use super::{
-    backup_and_write_json, backup_config_file, load_json_file, load_json_file_strict,
-    managed_rules_markdown, remove_legacy_rules_block, remove_managed_rules_file,
-    safe_write_json_file, write_managed_rules_file, AgentIntegration, DoctorCounters,
-    HealthcheckContext, InstallContext, InstallScope, RulesVariant,
-};
-
+use super::*;
 /// Filename of the managed rules file, colocated next to `opencode.json` and
 /// referenced from its `"instructions"` array.
 const INSTRUCTIONS_ENTRY: &str = "tokensave.md";
