+++ src/agents/integrations/pi.rs
@@ -15,10 +15,7 @@ use serde_json::json;
 
 use crate::errors::Result;
 
-use super::{
-    backup_and_write_json, backup_config_file, load_json_file, load_json_file_strict,
-    safe_write_json_file, AgentIntegration, DoctorCounters, HealthcheckContext, InstallContext,
-};
+use super::*;
 
 /// Pi coding agent.
 pub struct PiIntegration;
