+++ src/agents/integrations/vibe.rs
