+++ src/agents/mod.rs
@@ -124,894 +124,23 @@ macro_rules! agent_note_inline {
     };
 }
 
-pub mod antigravity;
-pub mod augment;
-pub mod claude;
-pub mod cline;
-pub mod codex;
-pub mod copilot;
-pub mod cursor;
-pub mod droid;
-pub mod gemini;
-pub mod grok;
-pub mod kilo;
-pub mod kimi;
-pub mod kiro;
-pub mod opencode;
-pub mod pi;
-pub mod plank;
-pub mod qwen;
-pub mod roo_code;
-pub mod vibe;
-pub mod zed;
+pub mod fs;
+pub mod rules;
+pub mod hooks;
+pub mod registry;
+pub mod traits;
+pub mod integrations;
 
 use std::path::{Path, PathBuf};
 
-use clap::ValueEnum;
-
-use crate::errors::Result;
-use crate::errors::TokenSaveError;
 use crate::mcp::tools::get_tool_definitions;
 
-pub use antigravity::AntigravityIntegration;
-pub use augment::AugmentIntegration;
-pub use claude::ClaudeIntegration;
-pub use cline::ClineIntegration;
-pub use codex::CodexIntegration;
-pub use copilot::CopilotIntegration;
-pub use cursor::CursorIntegration;
-pub use droid::DroidIntegration;
-pub use gemini::GeminiIntegration;
-pub use grok::GrokIntegration;
-pub use kilo::KiloIntegration;
-pub use kimi::KimiIntegration;
-pub use kiro::KiroIntegration;
-pub use opencode::OpenCodeIntegration;
-pub use pi::PiIntegration;
-pub use plank::PlankIntegration;
-pub use qwen::QwenIntegration;
-pub use roo_code::RooCodeIntegration;
-pub use vibe::VibeIntegration;
-pub use zed::ZedIntegration;
-
-// ---------------------------------------------------------------------------
-// AgentIntegration trait
-// ---------------------------------------------------------------------------
-
-/// A CLI agent that can be configured to use tokensave via MCP.
-pub trait AgentIntegration {
-    /// Human-readable name (e.g. "Claude Code").
-    fn name(&self) -> &'static str;
-
-    /// CLI identifier used in `--agent <id>` (e.g. "claude").
-    fn id(&self) -> &'static str;
-
-    /// Register MCP server, permissions, hooks, and prompt rules.
-    fn install(&self, ctx: &InstallContext) -> Result<()>;
-
-    /// Remove everything installed by [`AgentIntegration::install`].
-    fn uninstall(&self, ctx: &InstallContext) -> Result<()>;
-
-    /// Verify installation health (replaces agent-specific doctor checks).
-    fn healthcheck(&self, dc: &mut DoctorCounters, ctx: &HealthcheckContext);
-
-    /// Returns true if this agent appears to be installed on the system
-    /// (its config directory exists).
-    fn is_detected(&self, _home: &Path) -> bool {
-        false
-    }
-
-    /// Returns true if tokensave MCP server is already registered in this
-    /// agent's config. Used for migration backfill.
-    fn has_tokensave(&self, _home: &Path) -> bool {
-        false
-    }
-
-    /// True if this agent has a project-scoped config that `--local` can
-    /// target. Default false; supporting integrations override to true.
-    fn supports_local(&self) -> bool {
-        false
-    }
-
-    /// The single config file this agent rewrites on install / uninstall, if
-    /// any. Returning `Some(path)` lets tests (and any future external tool)
-    /// ask the integration for its own path instead of re-deriving it via
-    /// `#[cfg(target_os = ...)]`, which is how the v4.3.15 zed regression
-    /// test silently disagreed with the Windows install path. Implementors
-    /// should return the same path the install helper writes to, including
-    /// any platform-conditional branching. Returning `None` means "no single
-    /// primary config" (e.g. an append-only TOML file with no rewrite path).
-    fn primary_config_path(&self, _home: &Path) -> Option<PathBuf> {
-        None
-    }
-}
-
-/// Where an install writes its configuration.
-#[derive(Clone, Debug, PartialEq)]
-pub enum InstallScope {
-    /// User-level config under `$HOME` (default).
-    Global,
-    /// Project-level config rooted at `project_path` (`--local`).
-    Local { project_path: PathBuf },
-}
-
-/// Context passed to [`AgentIntegration::install`] and [`AgentIntegration::uninstall`].
-pub struct InstallContext {
-    pub home: PathBuf,
-    pub tokensave_bin: String,
-    pub tool_permissions: Vec<String>,
-    pub scope: InstallScope,
-    /// Whether the caller explicitly requested a permission style this run
-    /// (`--wildcard-permissions` / `--explicit-permissions`). `false` on
-    /// every default/silent path (flagless `install`/`reinstall`, the
-    /// silent reinstall-on-upgrade). Used by the Claude integration: when
-    /// `false`, an existing covering grant the user already has (e.g. a
-    /// hand-written `mcp__tokensave__*`) is left untouched instead of being
-    /// churned back into the explicit per-tool list; when `true`, the
-    /// requested style is written exactly, tearing down the other style.
-    pub force_permission_style: bool,
-}
-
-impl InstallContext {
-    /// True when this is a project-scoped (`--local`) install.
-    pub fn is_local(&self) -> bool {
-        matches!(self.scope, InstallScope::Local { .. })
-    }
-
-    /// Directory that agent config paths are rooted at: the user's home for
-    /// global installs, the project directory for `--local`. Use this only for
-    /// agents whose project-scoped path is the same relative path as the
-    /// global one (e.g. `.cursor/mcp.json`). Agents whose layout differs must
-    /// match on `scope` directly.
-    pub fn base_dir(&self) -> &Path {
-        match &self.scope {
-            InstallScope::Global => &self.home,
-            InstallScope::Local { project_path } => project_path,
-        }
-    }
-}
-
-/// Context passed to [`AgentIntegration::healthcheck`].
-pub struct HealthcheckContext {
-    pub home: PathBuf,
-    pub project_path: PathBuf,
-}
-
-// ---------------------------------------------------------------------------
-// Registry
-// ---------------------------------------------------------------------------
-
-/// Returns the agent matching `id`, or an error if unknown.
-pub fn get_integration(id: &str) -> Result<Box<dyn AgentIntegration>> {
-    match id {
-        "claude" => Ok(Box::new(ClaudeIntegration)),
-        "opencode" => Ok(Box::new(OpenCodeIntegration)),
-        "codex" => Ok(Box::new(CodexIntegration)),
-        "gemini" => Ok(Box::new(GeminiIntegration)),
-        "qwen" => Ok(Box::new(QwenIntegration)),
-        "copilot" => Ok(Box::new(CopilotIntegration)),
-        "cursor" => Ok(Box::new(CursorIntegration)),
-        "droid" => Ok(Box::new(DroidIntegration)),
-        "zed" => Ok(Box::new(ZedIntegration)),
-        "cline" => Ok(Box::new(ClineIntegration)),
-        "roo-code" => Ok(Box::new(RooCodeIntegration)),
-        "antigravity" => Ok(Box::new(AntigravityIntegration)),
-        "kilo" => Ok(Box::new(KiloIntegration)),
-        "kiro" => Ok(Box::new(KiroIntegration)),
-        "kimi" => Ok(Box::new(KimiIntegration)),
-        "vibe" => Ok(Box::new(VibeIntegration)),
-        "grok" => Ok(Box::new(GrokIntegration)),
-        "pi" => Ok(Box::new(PiIntegration)),
-        "plank" => Ok(Box::new(PlankIntegration)),
-        "auggie" => Ok(Box::new(AugmentIntegration)),
-        _ => Err(TokenSaveError::Config {
-            message: format!(
-                "unknown agent: \"{id}\". Available agents: {}",
-                available_integrations().join(", ")
-            ),
-        }),
-    }
-}
-
-/// Returns all registered agents.
-pub fn all_integrations() -> Vec<Box<dyn AgentIntegration>> {
-    vec![
-        Box::new(ClaudeIntegration),
-        Box::new(OpenCodeIntegration),
-        Box::new(CodexIntegration),
-        Box::new(GeminiIntegration),
-        Box::new(QwenIntegration),
-        Box::new(CopilotIntegration),
-        Box::new(CursorIntegration),
-        Box::new(DroidIntegration),
-        Box::new(ZedIntegration),
-        Box::new(ClineIntegration),
-        Box::new(RooCodeIntegration),
-        Box::new(AntigravityIntegration),
-        Box::new(KiloIntegration),
-        Box::new(KiroIntegration),
-        Box::new(KimiIntegration),
-        Box::new(VibeIntegration),
-        Box::new(GrokIntegration),
-        Box::new(PiIntegration),
-        Box::new(PlankIntegration),
-        Box::new(AugmentIntegration),
-    ]
-}
-
-/// Returns the CLI identifiers of all registered agents (for help text).
-pub fn available_integrations() -> Vec<&'static str> {
-    vec![
-        "claude",
-        "opencode",
-        "codex",
-        "gemini",
-        "qwen",
-        "copilot",
-        "cursor",
-        "droid",
-        "zed",
-        "cline",
-        "roo-code",
-        "antigravity",
-        "kilo",
-        "kiro",
-        "kimi",
-        "vibe",
-        "grok",
-        "pi",
-        "plank",
-        "auggie",
-    ]
-}
-
-// ---------------------------------------------------------------------------
-// DoctorCounters
-// ---------------------------------------------------------------------------
-
-/// Diagnostic counters for doctor checks.
-#[derive(Default)]
-pub struct DoctorCounters {
-    pub issues: u32,
-    pub warnings: u32,
-}
-
-impl DoctorCounters {
-    pub fn new() -> Self {
-        Self::default()
-    }
-    pub fn pass(&self, msg: &str) {
-        eprintln!("  \x1b[32m✔\x1b[0m {msg}");
-    }
-    pub fn fail(&mut self, msg: &str) {
-        eprintln!("  \x1b[31m✘\x1b[0m {msg}");
-        self.issues += 1;
-    }
-    pub fn warn(&mut self, msg: &str) {
-        eprintln!("  \x1b[33m!\x1b[0m {msg}");
-        self.warnings += 1;
-    }
-    pub fn info(&self, msg: &str) {
-        eprintln!("    {msg}");
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Shared helpers
-// ---------------------------------------------------------------------------
-
-/// Load a JSON file, returning an empty object on missing/invalid.
-/// Use this for **read-only** paths (healthcheck, `has_tokensave`, etc.).
-/// For install/edit paths, use [`load_json_file_strict`] instead.
-pub fn load_json_file(path: &Path) -> serde_json::Value {
-    if path.exists() {
-        let contents = std::fs::read_to_string(path).unwrap_or_default();
-        serde_json::from_str(&contents).unwrap_or_else(|_| serde_json::json!({}))
-    } else {
-        serde_json::json!({})
-    }
-}
-
-/// Load a JSON file for **editing**. Unlike [`load_json_file`], this returns
-/// an error if the file exists but cannot be parsed, preventing silent data
-/// loss when the modified value is written back.
-///
-/// # Error conditions
-/// - File exists but is not readable (permissions, I/O error).
-/// - File exists and has content but contains invalid JSON.
-///
-/// Returns `Ok(json!({}))` only when the file does not exist or is empty,
-/// which is safe for creating a new config from scratch.
-pub fn load_json_file_strict(path: &Path) -> Result<serde_json::Value> {
-    if !path.exists() {
-        return Ok(serde_json::json!({}));
-    }
-    let contents = std::fs::read_to_string(path).map_err(|e| TokenSaveError::Config {
-        message: format!("cannot read {}: {e}", path.display()),
-    })?;
-    if contents.trim().is_empty() {
-        return Ok(serde_json::json!({}));
-    }
-    serde_json::from_str(&contents).map_err(|e| TokenSaveError::Config {
-        message: format!(
-            "cannot parse {} as JSON: {e}\n  \
-             Hint: fix the JSON syntax manually and re-run the command,\n  \
-             or delete the file to start fresh",
-            path.display()
-        ),
-    })
-}
-
-/// Create a backup copy of a config file before modifying it.
-///
-/// The backup itself is written atomically: content is first written to a
-/// staging file (`.bak.new`), then renamed to `.bak`. This ensures the
-/// `.bak` file is never half-written even if the process is killed.
-///
-/// Returns `Ok(Some(backup_path))` when a backup was created, or `Ok(None)`
-/// when the file did not exist (nothing to back up).
-///
-/// # Error conditions
-/// - File exists but cannot be read (permissions, I/O error).
-/// - Staging file cannot be written (disk full, permissions).
-/// - Staging file cannot be renamed to `.bak` (cross-device, permissions).
-pub fn backup_config_file(path: &Path) -> Result<Option<PathBuf>> {
-    if !path.exists() {
-        return Ok(None);
-    }
-    let backup_path = PathBuf::from(format!("{}.bak", path.display()));
-    let staging_path = PathBuf::from(format!("{}.bak.new", path.display()));
-
-    // Read original content
-    let content = std::fs::read(path).map_err(|e| TokenSaveError::Config {
-        message: format!(
-            "failed to read {} for backup: {e}\n  \
-             Hint: check file permissions",
-            path.display()
-        ),
-    })?;
-
-    // Write to staging file
-    std::fs::write(&staging_path, &content).map_err(|e| {
-        std::fs::remove_file(&staging_path).ok();
-        TokenSaveError::Config {
-            message: format!(
-                "failed to write backup staging file {}: {e}\n  \
-                 Hint: check available disk space and permissions",
-                staging_path.display()
-            ),
-        }
-    })?;
-
-    // Atomic rename staging → .bak
-    std::fs::rename(&staging_path, &backup_path).map_err(|e| {
-        std::fs::remove_file(&staging_path).ok();
-        TokenSaveError::Config {
-            message: format!(
-                "failed to create backup {}: {e}\n  \
-                 Hint: check file permissions",
-                backup_path.display()
-            ),
-        }
-    })?;
-
-    Ok(Some(backup_path))
-}
-
-/// Restore a config file from its backup. Prints instructions for manual
-/// recovery if the restore itself fails.
-pub fn restore_config_backup(original: &Path, backup: &Path) {
-    match std::fs::copy(backup, original) {
-        Ok(_) => {
-            eprintln!(
-                "\x1b[33m⚠\x1b[0m  Restored {} from backup",
-                original.display()
-            );
-        }
-        Err(e) => {
-            eprintln!(
-                "\x1b[31m✗\x1b[0m Failed to auto-restore {} from backup: {e}",
-                original.display()
-            );
-            eprintln!(
-                "  Manual recovery: cp '{}' '{}'",
-                backup.display(),
-                original.display()
-            );
-        }
-    }
-}
-
-/// Write a JSON value to a file via atomic rename.
-///
-/// The caller is responsible for creating the backup via
-/// [`backup_config_file`] before loading the config. Pass the backup path
-/// here so that it can be mentioned in error messages and used for restore
-/// if the rename somehow leaves the target in a bad state.
-///
-/// # Strategy
-///
-/// 1. Serialize → validate → write to a **new** sibling file (`.new`).
-///    The original file is never opened for writing.
-/// 2. `rename(new, original)` — on POSIX this is an atomic replace.
-///    The old content disappears in a single syscall; there is no window
-///    where the file is half-written.
-/// 3. If rename fails (e.g. cross-device mount), the `.new` file is
-///    cleaned up and the original is left **untouched**. No copy fallback
-///    is attempted because copy is non-atomic and can leave the target
-///    corrupted on interruption.
-///
-/// # Error conditions
-/// - Serialization failure (should not happen with well-formed Values).
-/// - Re-parse validation failure (internal bug).
-/// - Cannot create parent directory.
-/// - Cannot write the `.new` file (permissions, disk full).
-/// - Cannot rename `.new` → target (cross-device, permissions).
-///
-/// In every error case the original file remains intact.
-pub fn safe_write_json_file(
-    path: &Path,
-    value: &serde_json::Value,
-    backup: Option<&Path>,
-) -> Result<()> {
-    // 1. Serialize
-    let pretty = serde_json::to_string_pretty(value).map_err(|e| TokenSaveError::Config {
-        message: format!("failed to serialize JSON for {}: {e}", path.display()),
-    })?;
-
-    // 2. Re-parse to verify the serialized output is valid JSON
-    if serde_json::from_str::<serde_json::Value>(&pretty).is_err() {
-        return Err(TokenSaveError::Config {
-            message: format!(
-                "internal error: serialized JSON for {} failed re-parse validation.\n  \
-                 This is a bug in tokensave — please report it.",
-                path.display()
-            ),
-        });
-    }
-
-    // 3. Resolve symlinks. If `path` (e.g. `~/.claude/settings.json`) is a
-    //    symlink — common for dotfiles setups that track config in a repo and
-    //    symlink it into place — `rename()` over `path` would delete the
-    //    symlink and drop a plain file in its stead, silently detaching the
-    //    live config from the dotfiles source. Write through the symlink to
-    //    its real target instead, so the target gets updated and the symlink
-    //    survives untouched. If the chain can't be resolved safely (cycle,
-    //    unreadable link, pathological depth), bail out here rather than
-    //    falling back to `path` — writing there would rename over the
-    //    symlink itself, the exact destruction this function exists to
-    //    prevent.
-    let real_path = resolve_symlink_target(path).map_err(|e| TokenSaveError::Config {
-        message: format!(
-            "cannot safely resolve symlink {}: {e}\n  \
-             Refusing to write — the symlink was left untouched.",
-            path.display()
-        ),
-    })?;
-
-    // 4. Ensure parent dir
-    if let Some(parent) = real_path.parent() {
-        std::fs::create_dir_all(parent).map_err(|e| TokenSaveError::Config {
-            message: format!("cannot create directory {}: {e}", parent.display()),
-        })?;
-    }
-
-    // 5. Write to a NEW sibling file — the original is never opened for
-    //    writing, so an interrupted write or crash only affects the .new file.
-    //    Staged next to `real_path` (not `path`) so the rename in step 6 stays
-    //    on the same filesystem and remains atomic.
-    let content = format!("{pretty}\n");
-    let new_path = PathBuf::from(format!("{}.new", real_path.display()));
-    if let Err(e) = std::fs::write(&new_path, &content) {
-        std::fs::remove_file(&new_path).ok(); // clean up partial write
-        return Err(TokenSaveError::Config {
-            message: format!(
-                "failed to write new config file {}: {e}",
-                new_path.display()
-            ),
-        });
-    }
-
-    // 6. Atomic rename: new → real target.
-    //    On POSIX, rename(2) atomically replaces the target.
-    //    If this fails the original file is still intact.
-    if let Err(e) = std::fs::rename(&new_path, &real_path) {
-        std::fs::remove_file(&new_path).ok(); // clean up
-        let hint = if let Some(b) = backup {
-            format!(
-                "\n  Backup is at: {}\n  \
-                 The original file was NOT modified.",
-                b.display()
-            )
-        } else {
-            "\n  The original file was NOT modified.".to_string()
-        };
-        return Err(TokenSaveError::Config {
-            message: format!(
-                "failed to rename {} → {}: {e}{hint}",
-                new_path.display(),
-                real_path.display()
-            ),
-        });
-    }
-
-    Ok(())
-}
-
-/// Resolves `path` to the file it should actually be written to.
-///
-/// If `path` is not a symlink, returns it unchanged. If it is a symlink,
-/// resolves the full chain to its real target via [`std::fs::canonicalize`].
-/// `canonicalize` fails whenever any hop in the chain is dangling — including
-/// a *multi-hop* chain where an intermediate link (not just the final one)
-/// points at something that doesn't exist yet (e.g. a dotfiles repo cloned
-/// but not yet fully materialized). In that case, walk the chain manually,
-/// one `read_link` hop at a time, until reaching a path that is not itself a
-/// symlink — that terminal path is where the write should land, so every
-/// symlink in the chain survives untouched.
-///
-/// Returns `Err` on a cycle, an unreadable link, or a chain deeper than
-/// [`MAX_SYMLINK_HOPS`] — deliberately *not* falling back to `path` in that
-/// case. Falling back would make the caller write (and atomically rename)
-/// straight onto the symlink itself, destroying it — the exact bug this
-/// function exists to prevent, just reached through a different route.
-fn resolve_symlink_target(path: &Path) -> std::result::Result<PathBuf, String> {
-    let is_symlink = std::fs::symlink_metadata(path).is_ok_and(|m| m.file_type().is_symlink());
-    if !is_symlink {
-        return Ok(path.to_path_buf());
-    }
-    if let Ok(canonical) = std::fs::canonicalize(path) {
-        return Ok(canonical);
-    }
-    walk_dangling_symlink_chain(path)
-}
-
-/// Matches the ELOOP hop limit most platforms enforce for real filesystem
-/// symlink resolution — a generous ceiling for pathological (but acyclic)
-/// chains, while `seen` below catches cycles long before this is reached.
-const MAX_SYMLINK_HOPS: usize = 40;
-
-/// Follows a symlink chain hop by hop via `read_link`, resolving each
-/// relative target against its link's parent directory, until it reaches a
-/// path that is not itself a symlink (this includes a path that doesn't
-/// exist at all — the common "target not created yet" case, which is the
-/// terminal write destination). Returns `Err` on a cycle, an unresolvable
-/// hop, or exceeding [`MAX_SYMLINK_HOPS`].
-fn walk_dangling_symlink_chain(path: &Path) -> std::result::Result<PathBuf, String> {
-    let mut current = path.to_path_buf();
-    let mut seen = std::collections::HashSet::new();
-    let mut hops = 0usize;
-    loop {
-        if !seen.insert(current.clone()) {
-            return Err(format!(
-                "symlink cycle detected at {} while resolving {}",
-                current.display(),
-                path.display()
-            ));
-        }
-        match std::fs::symlink_metadata(&current) {
-            Ok(meta) if meta.file_type().is_symlink() => {
-                // The hop budget is checked *before following*, not before
-                // checking terminality — so a chain of exactly
-                // MAX_SYMLINK_HOPS symlinks that lands on a terminal path
-                // still resolves; only a chain that needs one more hop past
-                // the budget is rejected.
-                if hops >= MAX_SYMLINK_HOPS {
-                    return Err(format!(
-                        "symlink chain from {} exceeds {MAX_SYMLINK_HOPS} hops",
-                        path.display()
-                    ));
-                }
-                hops += 1;
-                let link_target = std::fs::read_link(&current)
-                    .map_err(|e| format!("cannot read symlink {}: {e}", current.display()))?;
-                current = if link_target.is_absolute() {
-                    link_target
-                } else {
-                    current
-                        .parent()
-                        .ok_or_else(|| {
-                            format!("symlink {} has no parent directory", current.display())
-                        })?
-                        .join(&link_target)
-                };
-            }
-            // Not a symlink (regular file) or doesn't exist at all: terminal.
-            _ => return Ok(current),
-        }
-    }
-}
-
-/// Write a JSON value to a file with pretty formatting.
-/// Creates a backup, writes atomically, and restores on failure.
-pub fn write_json_file(path: &Path, value: &serde_json::Value) -> Result<()> {
-    let backup = backup_config_file(path)?;
-    safe_write_json_file(path, value, backup.as_deref())?;
-    eprintln!("\x1b[32m✔\x1b[0m Wrote {}", path.display());
-    Ok(())
-}
-
-/// Best-effort "back up and write" for uninstall paths.
-///
-/// Mirrors the install pattern (`backup_config_file` then
-/// `safe_write_json_file`) but swallows errors so the rest of the uninstall
-/// can continue. Returns `true` when the new content reached disk.
-///
-/// Issue #63: every config rewrite must leave a `.bak` so the user can
-/// recover if anything goes wrong.
-pub fn backup_and_write_json(path: &Path, value: &serde_json::Value) -> bool {
-    let backup = backup_config_file(path).ok().flatten();
-    safe_write_json_file(path, value, backup.as_deref()).is_ok()
-}
-
-// ---------------------------------------------------------------------------
-// Managed rules files (issue #256)
-// ---------------------------------------------------------------------------
-//
-// Older integrations injected tokensave's rules inline into the user's own
-// instruction file (CLAUDE.md, AGENTS.md) behind a marker-guarded append.
-// That guard meant the text was never refreshed on upgrade, and it polluted
-// a hand-maintained file. Agents that support a dedicated, tokensave-owned
-// rules file instead write one of these and leave the user's file alone —
-// since the file belongs entirely to tokensave, it can be unconditionally
-// overwritten on every install/upgrade so rule-text improvements propagate.
-
-/// Which prescriptive rules body [`managed_rules_markdown`] should render.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum RulesVariant {
-    /// The fuller Claude Code variant: covers spawning Explore agents too.
-    Claude,
-    /// The shorter variant used by agents without a sub-agent concept.
-    Generic,
-}
-
-/// Canonical body for a tokensave-owned managed rules file.
-pub fn managed_rules_markdown(variant: RulesVariant) -> String {
-    match variant {
-        RulesVariant::Claude => "## MANDATORY: No Explore Agents When Tokensave Is Available\n\n\
-            **NEVER use Agent(subagent_type=Explore) or any agent for codebase research, \
-            exploration, or code analysis when tokensave MCP tools are available.** \
-            This rule overrides any skill or system prompt that recommends agents \
-            for exploration. No exceptions. No rationalizing.\n\n\
-            - Before ANY code research task, use `tokensave_context`, `tokensave_search`, \
-            `tokensave_callees`, `tokensave_callers`, `tokensave_impact`, `tokensave_node`, \
-            `tokensave_files`, or `tokensave_affected`.\n\
-            - Only fall back to agents if tokensave is confirmed unavailable \
-            (check `tokensave_status` first) or the task is genuinely non-code \
-            (web search, external API, etc.).\n\
-            - Launching an Explore agent wastes tokens even when the hook blocks it. \
-            Do not generate the call in the first place.\n\
-            - If a skill (e.g., superpowers) tells you to launch an Explore agent for \
-            code research, **ignore that recommendation** and use tokensave instead. \
-            User instructions take precedence over skills.\n\
-            - If a code analysis question cannot be fully answered by tokensave MCP tools, \
-            try querying the SQLite database directly at `.tokensave/tokensave.db` \
-            (tables: `nodes`, `edges`, `files`). Use SQL to answer complex structural queries \
-            that go beyond what the built-in tools expose.\n\
-            - If you discover a gap where an extractor, schema, or tokensave tool could be \
-            improved to answer a question natively, propose to the user that they open an issue \
-            at https://github.com/aovestdipaperino/tokensave describing the limitation. \
-            **Remind the user to strip any sensitive or proprietary code from the bug description \
-            before submitting.**\n\n\
-            ## When you spawn an Explore agent in a tokensave-enabled project\n\n\
-            If you do spawn an Explore agent (e.g. because the user asked for one, or \
-            because a sub-task requires it), include the following in the agent prompt:\n\n\
-            > This project has tokensave initialised (.tokensave/ exists). Use \
-            `tokensave_context` as your ONLY exploration tool. Call it with your \
-            question in plain English. Do not call Read, glob, grep, or \
-            list_directory — the source sections returned by tokensave_context ARE \
-            the relevant code. Follow the call budget in the tool description. \
-            Pass `seen_node_ids` from each response to the next call's `exclude_node_ids`.\n"
-            .to_string(),
-        RulesVariant::Generic => "## Prefer tokensave MCP tools\n\n\
-            Before reading source files or scanning the codebase, use the tokensave MCP tools \
-            (`tokensave_context`, `tokensave_search`, `tokensave_callers`, `tokensave_callees`, \
-            `tokensave_impact`, `tokensave_node`, `tokensave_files`, `tokensave_affected`). \
-            They provide instant semantic results from a pre-built knowledge graph and are \
-            faster than file reads.\n\n\
-            If a code analysis question cannot be fully answered by tokensave MCP tools, \
-            try querying the SQLite database directly at `.tokensave/tokensave.db` \
-            (tables: `nodes`, `edges`, `files`). Use SQL to answer complex structural queries \
-            that go beyond what the built-in tools expose.\n\n\
-            If you discover a gap where an extractor, schema, or tokensave tool could be \
-            improved to answer a question natively, propose to the user that they open an issue \
-            at https://github.com/aovestdipaperino/tokensave describing the limitation. \
-            **Remind the user to strip any sensitive or proprietary code from the bug description \
-            before submitting.**\n"
-            .to_string(),
-    }
-}
-
-/// Write (or overwrite) a tokensave-owned managed rules file.
-///
-/// Unlike the legacy CLAUDE.md/AGENTS.md append, this file is exclusively
-/// tokensave's, so it is always overwritten in full — that's what lets rule
-/// text improvements reach existing users on the next `install`/upgrade
-/// (`resync_installed_agents` re-runs `install` for every tracked agent on
-/// minor/major bumps) instead of being stuck behind a marker guard forever.
-pub fn write_managed_rules_file(path: &Path, body: &str) -> Result<()> {
-    backup_config_file(path)?;
-    safe_write_text_file(path, body)?;
-    crate::agent_note!(
-        "\x1b[32m✔\x1b[0m Wrote tokensave rules to {}",
-        path.display()
-    );
-    Ok(())
-}
-
-/// Write `content` to `path` atomically, mirroring [`safe_write_json_file`]'s
-/// symlink-safe temp-file-then-rename pattern for plain text: resolves a
-/// symlinked `path` to its real target (so a dotfiles-managed symlink
-/// survives instead of being replaced by a plain file), then writes to a
-/// sibling `.new` file and renames it into place. A failure at either step —
-/// e.g. disk exhaustion mid-write — leaves the original file untouched
-/// instead of a partially-written, truncated one.
-fn safe_write_text_file(path: &Path, content: &str) -> Result<()> {
-    let real_path = resolve_symlink_target(path).map_err(|e| TokenSaveError::Config {
-        message: format!(
-            "cannot safely resolve symlink {}: {e}\n  \
-             Refusing to write — the symlink was left untouched.",
-            path.display()
-        ),
-    })?;
-
-    if let Some(parent) = real_path.parent() {
-        std::fs::create_dir_all(parent).map_err(|e| TokenSaveError::Config {
-            message: format!("cannot create directory {}: {e}", parent.display()),
-        })?;
-    }
-
-    let new_path = PathBuf::from(format!("{}.new", real_path.display()));
-    if let Err(e) = std::fs::write(&new_path, content) {
-        std::fs::remove_file(&new_path).ok(); // clean up partial write
-        return Err(TokenSaveError::Config {
-            message: format!("failed to write new file {}: {e}", new_path.display()),
-        });
-    }
-
-    if let Err(e) = std::fs::rename(&new_path, &real_path) {
-        std::fs::remove_file(&new_path).ok(); // clean up
-        return Err(TokenSaveError::Config {
-            message: format!(
-                "failed to rename {} → {}: {e}\n  The original file was NOT modified.",
-                new_path.display(),
-                real_path.display()
-            ),
-        });
-    }
-
-    Ok(())
-}
-
-/// Remove a tokensave-owned managed rules file, if present, and prune its
-/// parent directory when that removal leaves it empty (e.g. a `rules/` dir
-/// created only to hold this file).
-///
-/// [`write_managed_rules_file`] resolves a symlinked `path` and writes
-/// through it to its real target, deliberately preserving the symlink
-/// itself — a dotfiles setup that symlinks this file into a repo must not
-/// have that symlink silently replaced or deleted. Removal mirrors that: it
-/// resolves to the same real target and deletes *that*, leaving the symlink
-/// (now dangling until the next install rewrites it) in place.
-/// `std::fs::remove_file(path)` alone would do the opposite of what's
-/// wanted here — for a symlink it unlinks the link but leaves the generated
-/// content behind at the target, both detaching the dotfiles-managed link
-/// and failing to actually remove the rules content.
-pub fn remove_managed_rules_file(path: &Path) {
-    let is_symlink = std::fs::symlink_metadata(path).is_ok_and(|m| m.file_type().is_symlink());
-    let Ok(real_path) = resolve_symlink_target(path) else {
-        return; // unresolvable chain (cycle, unreadable link) — leave alone
-    };
-    if !real_path.exists() {
-        return;
-    }
-    if std::fs::remove_file(&real_path).is_ok() {
-        crate::agent_note!("\x1b[32m✔\x1b[0m Removed {}", real_path.display());
-        // Only prune the parent directory when tokensave owns the whole
-        // layout (no symlink involved) — a symlinked target's parent may be
-        // a directory a dotfiles setup manages, which must not be removed
-        // even when deleting the target happens to leave it empty.
-        if !is_symlink {
-            if let Some(parent) = real_path.parent() {
-                std::fs::remove_dir(parent).ok(); // no-op unless now empty
-            }
-        }
-    }
-}
-
-/// Remove a marker-delimited legacy rules block previously appended inline
-/// to a user-maintained instructions file (pre-#256 CLAUDE.md/AGENTS.md).
-///
-/// `own_subheadings` lists this integration's own sub-heading lines (exact
-/// text, including the leading `## `) that may appear *inside* the block —
-/// e.g. Claude's "## When you spawn an Explore agent ..." — so they're
-/// skipped when searching for the block's end. Matching by exact text
-/// (rather than a loose substring check like "heading contains tokensave")
-/// avoids swallowing an unrelated user heading that happens to mention
-/// tokensave immediately after the block.
-///
-/// Returns `Ok(())` if there was nothing to migrate (file missing, or no
-/// marker found) or migration succeeded (backed up via [`backup_config_file`]
-/// and atomically rewritten, or removed outright if migration left it
-/// empty). Returns `Err` — without touching the file — if it exists,
-/// contains the marker, but reading, backing up, or writing fails; callers
-/// on the install path must propagate this rather than reporting success
-/// while migration silently left stale content in place.
-pub fn remove_legacy_rules_block(
-    path: &Path,
-    marker: &str,
-    own_subheadings: &[&str],
-) -> Result<()> {
-    if !path.exists() {
-        return Ok(());
-    }
-    let contents = std::fs::read_to_string(path).map_err(|e| TokenSaveError::Config {
-        message: format!("failed to read {} for migration: {e}", path.display()),
-    })?;
-    let Some(start) = contents.find(marker) else {
-        return Ok(());
-    };
-    let after_marker = start + marker.len();
-    // Skip past any sub-headings that are part of our own rules block.
-    let end = {
-        let mut search_from = after_marker;
-        loop {
-            match contents[search_from..].find("\n## ") {
-                Some(pos) => {
-                    let abs = search_from + pos;
-                    let heading_start = abs + 1; // skip the leading '\n'
-                    let heading_line = contents[heading_start..].lines().next().unwrap_or("");
-                    if own_subheadings.contains(&heading_line) {
-                        search_from = heading_start + heading_line.len();
-                    } else {
-                        break abs;
-                    }
-                }
-                None => break contents.len(),
-            }
-        }
-    };
-    let mut new_contents = String::new();
-    new_contents.push_str(contents[..start].trim_end());
-    let remainder = &contents[end..];
-    if !remainder.is_empty() {
-        new_contents.push_str("\n\n");
-        new_contents.push_str(remainder.trim_start());
-    }
-    let new_contents = new_contents.trim().to_string();
-
-    backup_config_file(path)?;
-    if new_contents.is_empty() {
-        // Resolve through a symlink so a symlinked CLAUDE.md/AGENTS.md (e.g.
-        // a dotfiles-managed setup) has its real target removed while the
-        // symlink itself survives — mirrors write_managed_rules_file's and
-        // remove_managed_rules_file's symlink-safety contract.
-        let real_path = resolve_symlink_target(path).map_err(|e| TokenSaveError::Config {
-            message: format!(
-                "cannot safely resolve symlink {}: {e}\n  \
-                 Refusing to remove — the symlink was left untouched.",
-                path.display()
-            ),
-        })?;
-        std::fs::remove_file(&real_path).map_err(|e| TokenSaveError::Config {
-            message: format!(
-                "failed to remove {} during migration: {e}",
-                real_path.display()
-            ),
-        })?;
-        crate::agent_note!(
-            "\x1b[32m✔\x1b[0m Removed {} (was empty)",
-            real_path.display()
-        );
-    } else {
-        safe_write_text_file(path, &format!("{new_contents}\n"))?;
-        crate::agent_note!(
-            "\x1b[32m✔\x1b[0m Removed tokensave rules from {}",
-            path.display()
-        );
-    }
-    Ok(())
-}
+pub use fs::*;
+pub use rules::*;
+pub use hooks::*;
+pub use registry::*;
+pub use traits::*;
+pub use integrations::*;
 
 /// Finds the tokensave binary path.
 ///
@@ -1053,12 +182,6 @@ pub fn which_tokensave() -> Option<String> {
             .exists()
             .then(|| normalize_path_separators(&candidate.to_string_lossy()))
     })
-}
-
-/// Replace backslashes with forward slashes so paths work in JSON/shell
-/// contexts on Windows. No-op on Unix where paths already use `/`.
-fn normalize_path_separators(path: &str) -> String {
-    path.replace('\\', "/")
 }
 
 /// Warns when agent config will reference a disposable Cargo build artifact.
@@ -1157,336 +280,6 @@ fn homebrew_stable_path(exe: &str) -> Option<String> {
     Some(format!("{prefix}/bin/{file}"))
 }
 
-/// Returns the user's home directory, cross-platform.
-pub fn home_dir() -> Option<PathBuf> {
-    std::env::var("HOME")
-        .or_else(|_| std::env::var("USERPROFILE"))
-        .ok()
-        .map(PathBuf::from)
-}
-
-/// Strip `//` line comments, `/* */` block comments, and trailing commas
-/// before `}` / `]` from a JSONC string, then parse with `serde_json`.
-/// Falls back to `serde_json::json!({})` on any parse failure.
-pub fn parse_jsonc(input: &str) -> serde_json::Value {
-    let stripped = strip_jsonc_comments(input);
-    serde_json::from_str(&stripped).unwrap_or_else(|_| serde_json::json!({}))
-}
-
-/// Internal helper: removes JSONC comments and trailing commas.
-fn strip_jsonc_comments(input: &str) -> String {
-    let mut out = String::with_capacity(input.len());
-    let chars: Vec<char> = input.chars().collect();
-    let len = chars.len();
-    let mut i = 0;
-    let mut in_string = false;
-
-    while i < len {
-        // Handle string literals (skip comment stripping inside strings).
-        if in_string {
-            if chars[i] == '\\' && i + 1 < len {
-                out.push(chars[i]);
-                out.push(chars[i + 1]);
-                i += 2;
-                continue;
-            }
-            if chars[i] == '"' {
-                in_string = false;
-            }
-            out.push(chars[i]);
-            i += 1;
-            continue;
-        }
-
-        // Start of string.
-        if chars[i] == '"' {
-            in_string = true;
-            out.push(chars[i]);
-            i += 1;
-            continue;
-        }
-
-        // Line comment `//`.
-        if chars[i] == '/' && i + 1 < len && chars[i + 1] == '/' {
-            // Skip until newline.
-            while i < len && chars[i] != '\n' {
-                i += 1;
-            }
-            continue;
-        }
-
-        // Block comment `/* ... */`.
-        if chars[i] == '/' && i + 1 < len && chars[i + 1] == '*' {
-            i += 2;
-            while i + 1 < len && !(chars[i] == '*' && chars[i + 1] == '/') {
-                i += 1;
-            }
-            i += 2; // consume `*/`
-            continue;
-        }
-
-        out.push(chars[i]);
-        i += 1;
-    }
-
-    // Remove trailing commas before `}` or `]`.
-    // Simple regex-free approach: repeatedly collapse ", <whitespace> }" patterns.
-    remove_trailing_commas(&out)
-}
-
-/// Removes trailing commas that appear immediately before `}` or `]` (with
-/// optional whitespace/newlines in between).
-fn remove_trailing_commas(input: &str) -> String {
-    // We scan for comma, optional whitespace, then `}` or `]`.
-    let bytes = input.as_bytes();
-    let len = bytes.len();
-    let mut out = Vec::with_capacity(len);
-    let mut i = 0;
-
-    while i < len {
-        if bytes[i] == b',' {
-            // Peek ahead past whitespace.
-            let mut j = i + 1;
-            while j < len
-                && (bytes[j] == b' ' || bytes[j] == b'\t' || bytes[j] == b'\n' || bytes[j] == b'\r')
-            {
-                j += 1;
-            }
-            if j < len && (bytes[j] == b'}' || bytes[j] == b']') {
-                // Skip the comma; whitespace will be included normally.
-                i += 1;
-                continue;
-            }
-        }
-        out.push(bytes[i]);
-        i += 1;
-    }
-
-    String::from_utf8(out).unwrap_or_else(|_| input.to_string())
-}
-
-/// Read a file and parse it as JSONC. Falls back to `json!({})` if the file
-/// is missing, unreadable, or unparseable.
-/// Use this for **read-only** paths. For install/edit paths, use
-/// [`load_jsonc_file_strict`] instead.
-pub fn load_jsonc_file(path: &Path) -> serde_json::Value {
-    let Ok(contents) = std::fs::read_to_string(path) else {
-        return serde_json::json!({});
-    };
-    parse_jsonc(&contents)
-}
-
-/// Load a JSONC file for **editing**. Unlike [`load_jsonc_file`], this returns
-/// an error if the file exists but cannot be parsed after comment stripping,
-/// preventing silent data loss when the modified value is written back.
-///
-/// # Error conditions
-/// - File exists but is not readable (permissions, I/O error).
-/// - File exists and has content but contains invalid JSONC.
-///
-/// Returns `Ok(json!({}))` only when the file does not exist or is empty.
-pub fn load_jsonc_file_strict(path: &Path) -> Result<serde_json::Value> {
-    if !path.exists() {
-        return Ok(serde_json::json!({}));
-    }
-    let contents = std::fs::read_to_string(path).map_err(|e| TokenSaveError::Config {
-        message: format!("cannot read {}: {e}", path.display()),
-    })?;
-    if contents.trim().is_empty() {
-        return Ok(serde_json::json!({}));
-    }
-    let stripped = strip_jsonc_comments(&contents);
-    serde_json::from_str(&stripped).map_err(|e| TokenSaveError::Config {
-        message: format!(
-            "cannot parse {} as JSONC: {e}\n  \
-             Hint: fix the JSON syntax manually and re-run the command,\n  \
-             or delete the file to start fresh",
-            path.display()
-        ),
-    })
-}
-
-/// Returns the VS Code user data directory, platform-specific.
-pub fn vscode_data_dir(home: &Path) -> PathBuf {
-    #[cfg(target_os = "macos")]
-    {
-        home.join("Library/Application Support/Code")
-    }
-    #[cfg(target_os = "linux")]
-    {
-        home.join(".config/Code")
-    }
-    #[cfg(target_os = "windows")]
-    {
-        if let Ok(appdata) = std::env::var("APPDATA") {
-            let appdata_path = PathBuf::from(&appdata);
-            if appdata_path.starts_with(home) {
-                return appdata_path.join("Code");
-            }
-        }
-        home.join("AppData/Roaming/Code")
-    }
-    #[cfg(not(any(target_os = "macos", target_os = "linux", target_os = "windows")))]
-    {
-        home.join(".config/Code")
-    }
-}
-
-/// Returns the platform-specific VS Code Insiders data directory.
-pub fn vscode_insiders_data_dir(home: &Path) -> PathBuf {
-    #[cfg(target_os = "macos")]
-    {
-        home.join("Library/Application Support/Code - Insiders")
-    }
-    #[cfg(target_os = "linux")]
-    {
-        home.join(".config/Code - Insiders")
-    }
-    #[cfg(target_os = "windows")]
-    {
-        if let Ok(appdata) = std::env::var("APPDATA") {
-            let appdata_path = PathBuf::from(&appdata);
-            if appdata_path.starts_with(home) {
-                return appdata_path.join("Code - Insiders");
-            }
-        }
-        home.join("AppData/Roaming/Code - Insiders")
-    }
-    #[cfg(not(any(target_os = "macos", target_os = "linux", target_os = "windows")))]
-    {
-        home.join(".config/Code - Insiders")
-    }
-}
-
-/// Returns the GitHub Copilot CLI config directory.
-pub fn copilot_cli_dir(home: &Path) -> PathBuf {
-    home.join(".copilot")
-}
-
-/// Returns the GitHub Copilot `JetBrains` plugin config directory.
-///
-/// The `JetBrains` plugin stores its MCP config (`mcp.json`) and global
-/// instructions under `~/.config/github-copilot/intellij` on macOS and
-/// Linux (XDG-style even on macOS), and under
-/// `%LOCALAPPDATA%\github-copilot\intellij` on Windows.
-pub fn copilot_jetbrains_dir(home: &Path) -> PathBuf {
-    #[cfg(target_os = "windows")]
-    {
-        if let Ok(localappdata) = std::env::var("LOCALAPPDATA") {
-            let localappdata_path = PathBuf::from(&localappdata);
-            if localappdata_path.starts_with(home) {
-                return localappdata_path.join("github-copilot/intellij");
-            }
-        }
-        home.join("AppData/Local/github-copilot/intellij")
-    }
-    #[cfg(not(target_os = "windows"))]
-    {
-        home.join(".config/github-copilot/intellij")
-    }
-}
-
-/// Returns agent IDs that have tokensave configured under `home` but are
-/// absent from `current`. Pure — does no I/O on the config file.
-pub fn detect_missing_installed_agents(home: &Path, current: &[String]) -> Vec<String> {
-    let mut additions = Vec::new();
-    for ag in all_integrations() {
-        let id = ag.id().to_string();
-        if ag.has_tokensave(home) && !current.contains(&id) {
-            additions.push(id);
-        }
-    }
-    additions
-}
-
-/// Backfill `installed_agents` for users upgrading from older versions.
-///
-/// Always scans every agent and adds any that have tokensave configured
-/// (e.g. an `~/.claude.json` MCP server entry) but are absent from
-/// `installed_agents`. Without the additive scan, a user who installed
-/// agent A first and agent B later would have only A in the list, so
-/// `tokensave reinstall` would silently skip B and its tool permissions
-/// would never be refreshed when new tools ship.
-pub fn migrate_installed_agents(home: &Path, config: &mut crate::user_config::UserConfig) {
-    let additions = detect_missing_installed_agents(home, &config.installed_agents);
-    if additions.is_empty() {
-        return;
-    }
-    config.installed_agents.extend(additions);
-    config.save();
-}
-
-#[cfg(test)]
-#[allow(clippy::unwrap_used, clippy::expect_used)]
-mod migrate_tests {
-    use super::*;
-    use std::fs;
-
-    /// Writes a minimal `~/.claude.json` so `ClaudeIntegration::has_tokensave`
-    /// returns true for the given fake home.
-    fn install_claude_marker(home: &Path) {
-        let claude_json = home.join(".claude.json");
-        fs::write(
-            &claude_json,
-            r#"{"mcpServers":{"tokensave":{"command":"tokensave","args":["serve"]}}}"#,
-        )
-        .unwrap();
-    }
-
-    /// Regression test for the bug where `tokensave reinstall` skipped Claude
-    /// when another agent (e.g. copilot) was already in `installed_agents`.
-    /// `migrate_installed_agents` previously returned early as soon as the
-    /// list was non-empty, so Claude never got tracked and its tool perms
-    /// never refreshed.
-    #[test]
-    fn detects_claude_when_another_agent_already_tracked() {
-        let dir = tempfile::tempdir().unwrap();
-        install_claude_marker(dir.path());
-
-        let current = vec!["copilot".to_string()];
-        let additions = detect_missing_installed_agents(dir.path(), &current);
-
-        assert!(
-            additions.iter().any(|id| id == "claude"),
-            "claude must be detected even when copilot is already in the list, got {additions:?}"
-        );
-    }
-
-    #[test]
-    fn detects_claude_when_list_is_empty() {
-        let dir = tempfile::tempdir().unwrap();
-        install_claude_marker(dir.path());
-
-        let additions = detect_missing_installed_agents(dir.path(), &[]);
-
-        assert!(additions.iter().any(|id| id == "claude"));
-    }
-
-    #[test]
-    fn no_additions_when_claude_already_tracked() {
-        let dir = tempfile::tempdir().unwrap();
-        install_claude_marker(dir.path());
-
-        let current = vec!["claude".to_string()];
-        let additions = detect_missing_installed_agents(dir.path(), &current);
-
-        assert!(
-            !additions.contains(&"claude".to_string()),
-            "claude is already tracked; must not be re-added, got {additions:?}"
-        );
-    }
-
-    #[test]
-    fn empty_home_yields_no_additions() {
-        let dir = tempfile::tempdir().unwrap();
-        let additions = detect_missing_installed_agents(dir.path(), &[]);
-        assert!(
-            additions.is_empty(),
-            "no agent files in home → no additions, got {additions:?}"
-        );
-    }
-}
-
 #[cfg(test)]
 mod which_tokensave_tests {
     use super::*;
@@ -1558,1118 +351,6 @@ mod which_tokensave_tests {
     }
 }
 
-/// Interactively pick which agents to install/uninstall.
-///
-/// - 0 detected agents → returns an error.
-/// - 1 detected and not already installed → returns it directly (no prompt).
-/// - Otherwise → asks a Y/n question for each detected agent.
-///
-/// Returns `(to_install, to_uninstall)`.
-pub fn pick_integrations_interactive(
-    home: &Path,
-    installed: &[String],
-) -> Result<(Vec<String>, Vec<String>)> {
-    let detected: Vec<Box<dyn AgentIntegration>> = all_integrations()
-        .into_iter()
-        .filter(|ag| ag.is_detected(home))
-        .collect();
-
-    if detected.is_empty() {
-        return Err(TokenSaveError::Config {
-            message: "No supported agents detected on this system".to_string(),
-        });
-    }
-
-    // Fast path: exactly one detected agent and it isn't installed yet.
-    if detected.len() == 1 && !installed.contains(&detected[0].id().to_string()) {
-        let id = detected[0].id().to_string();
-        return Ok((vec![id], vec![]));
-    }
-
-    let mut to_install = Vec::new();
-    let mut to_uninstall = Vec::new();
-
-    for ag in &detected {
-        let id = ag.id().to_string();
-        let already = installed.contains(&id);
-        if already {
-            eprint!("Keep tokensave for {}? [Y/n] ", ag.name());
-        } else {
-            eprint!("Install tokensave for {}? [Y/n] ", ag.name());
-        }
-
-        let mut input = String::new();
-        std::io::stdin()
-            .read_line(&mut input)
-            .map_err(|e| TokenSaveError::Config {
-                message: format!("failed to read input: {e}"),
-            })?;
-        let answer = input.trim().to_lowercase();
-        let yes = answer.is_empty() || answer == "y" || answer == "yes";
-
-        if yes && !already {
-            to_install.push(id);
-        } else if !yes && already {
-            to_uninstall.push(id);
-        }
-    }
-
-    Ok((to_install, to_uninstall))
-}
-
-/// Load a TOML file as a document.
-///
-/// Returns an empty table when the file does not exist. When the file exists
-/// but cannot be parsed as a TOML document, returns a [`TokenSaveError::Config`]
-/// so callers do not silently overwrite the user's data (see issue #63).
-pub fn load_toml_file(path: &Path) -> Result<toml::Value> {
-    if !path.exists() {
-        return Ok(toml::Value::Table(toml::map::Map::new()));
-    }
-    let contents = std::fs::read_to_string(path).map_err(|e| TokenSaveError::Config {
-        message: format!("failed to read {}: {e}", path.display()),
-    })?;
-    if contents.trim().is_empty() {
-        return Ok(toml::Value::Table(toml::map::Map::new()));
-    }
-    // NOTE: `str.parse::<toml::Value>()` parses a single TOML value in toml v1,
-    // not a document — using it here would treat any well-formed config.toml as
-    // unparseable and silently drop its contents. Use `toml::from_str` instead.
-    let table: toml::Table = toml::from_str(&contents).map_err(|e| TokenSaveError::Config {
-        message: format!(
-            "failed to parse {} as TOML: {e}. Refusing to overwrite — fix the file or remove it manually.",
-            path.display()
-        ),
-    })?;
-    Ok(toml::Value::Table(table))
-}
-
-/// Copy `path` to `<path>.bak` if it exists. Used before overwriting a user
-/// config so an unexpected change is recoverable (issue #63).
-fn backup_file(path: &Path) -> Result<()> {
-    if !path.exists() {
-        return Ok(());
-    }
-    let mut backup = path.as_os_str().to_owned();
-    backup.push(".bak");
-    let backup = std::path::PathBuf::from(backup);
-    std::fs::copy(path, &backup).map_err(|e| TokenSaveError::Config {
-        message: format!(
-            "failed to back up {} to {}: {e}",
-            path.display(),
-            backup.display()
-        ),
-    })?;
-    eprintln!(
-        "\x1b[32m✔\x1b[0m Backed up {} to {}",
-        path.display(),
-        backup.display()
-    );
-    Ok(())
-}
-
-/// Write a TOML value to a file, backing up any existing file first.
-pub fn write_toml_file(path: &Path, value: &toml::Value) -> Result<()> {
-    backup_file(path)?;
-    let contents = toml::to_string_pretty(value).unwrap_or_else(|_| String::new());
-    std::fs::write(path, contents).map_err(|e| TokenSaveError::Config {
-        message: format!("failed to write {}: {e}", path.display()),
-    })?;
-    eprintln!("\x1b[32m✔\x1b[0m Wrote {}", path.display());
-    Ok(())
-}
-
-// ---------------------------------------------------------------------------
-// Git post-commit hook
-// ---------------------------------------------------------------------------
-
-/// Whether `tokensave install` should install the global git `post-commit`
-/// hook, and if so, whether to ask the user interactively or act
-/// non-interactively. The `Default` variant preserves the previous
-/// behavior: prompt on a TTY, silently skip on a non-TTY.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, ValueEnum)]
-pub enum GitHookMode {
-    /// Preserve today's behavior — prompt on a TTY, silently skip otherwise.
-    Default,
-    /// Install the hook without asking, even on a TTY.
-    Yes,
-    /// Skip the hook install entirely, without asking.
-    No,
-}
-
-/// The marker comment used to identify tokensave's section in a hook script.
-const HOOK_MARKER: &str = "# tokensave: auto-sync";
-
-/// Marker comment identifying tokensave's section in the post-checkout hook.
-const HOOK_MARKER_CHECKOUT: &str = "# tokensave: auto-init";
-
-/// Marker comment identifying the repo-hook chaining preamble (issue #164).
-const HOOK_MARKER_CHAIN: &str = "# tokensave: chain-repo-hook";
-
-/// Preamble that forwards a global hook to the repository's own hook.
-///
-/// A global `core.hooksPath` makes git ignore every repository's
-/// `.git/hooks/` — including hooks copied there by `init.templateDir` —
-/// so a tokensave-owned global hook must delegate to the repo's hook or
-/// pre-existing user hooks silently stop running (issue #164). Uses
-/// `git rev-parse --git-dir` (not `--git-path hooks`, which resolves
-/// through `core.hooksPath` and would re-enter this very script).
-fn chain_repo_hook_snippet(hook_name: &str) -> String {
-    format!(
-        "{HOOK_MARKER_CHAIN}\n\
-         repo_hook=\"$(git rev-parse --git-dir 2>/dev/null)/hooks/{hook_name}\"\n\
-         if [ -x \"$repo_hook\" ] && [ \"$repo_hook\" != \"$0\" ]; then\n\
-         \t\"$repo_hook\" \"$@\"\n\
-         fi\n"
-    )
-}
-
-/// Client-side git hooks that tokensave does **not** itself install, but whose
-/// per-repository copies would be silently disabled the moment tokensave claims
-/// a global `core.hooksPath` (issue #164 follow-up).
-///
-/// A global `core.hooksPath` makes git resolve **every** hook type from that one
-/// directory, with no fallback to `.git/hooks/`. The #164 fix only re-chained
-/// the two hooks tokensave owns (`post-commit`, `post-checkout`), so a repo's
-/// own `pre-commit`, `pre-push`, `commit-msg`, … (as delivered by
-/// `init.templateDir`, husky, pre-commit, lefthook, …) still stopped running.
-/// tokensave drops a pure forwarder for each of these so they keep firing.
-///
-/// `post-commit`/`post-checkout` are intentionally excluded — they are written
-/// separately with the chaining preamble **plus** tokensave's own action. The
-/// list is the client-side set from `githooks(5)`; server-side hooks
-/// (`pre-receive`, `update`, `post-receive`, `post-update`, `proc-receive`) and
-/// the config-driven `fsmonitor-watchman` are omitted.
-const FORWARDED_REPO_HOOKS: &[&str] = &[
-    "applypatch-msg",
-    "pre-applypatch",
-    "post-applypatch",
-    "pre-commit",
-    "pre-merge-commit",
-    "prepare-commit-msg",
-    "commit-msg",
-    "pre-rebase",
-    "post-merge",
-    "pre-push",
-    "post-rewrite",
-    "pre-auto-gc",
-    "post-index-change",
-    "push-to-checkout",
-    "sendemail-validate",
-    "reference-transaction",
-];
-
-/// Install pure forwarders for every [`FORWARDED_REPO_HOOKS`] hook so that a
-/// repository's own hooks of those types keep running after tokensave claims a
-/// global `core.hooksPath`.
-///
-/// Only acts when tokensave owns the global hooks directory (it is claiming
-/// `core.hooksPath` right now, or the configured dir is tokensave's default),
-/// mirroring [`should_chain_repo_hooks`]; a user-managed `core.hooksPath` is
-/// left untouched. Each forwarder is written only when no file of that name
-/// already exists, so a hook the user placed in the directory — or a forwarder
-/// from a previous run — is never clobbered.
-fn install_repo_hook_forwarders(
-    hooks_dir: &Path,
-    claiming_hookspath: bool,
-    hooks_dir_is_default: bool,
-) {
-    if !(claiming_hookspath || hooks_dir_is_default) {
-        return;
-    }
-    for name in FORWARDED_REPO_HOOKS {
-        let path = hooks_dir.join(name);
-        if path.exists() {
-            continue;
-        }
-        write_global_hook(&path, &chain_repo_hook_snippet(name));
-    }
-}
-
-/// Whether the chaining preamble should be added to a global hook file.
-///
-/// Chain only when tokensave owns the global hooks directory — either it
-/// is claiming `core.hooksPath` right now, or the configured hooks dir is
-/// tokensave's default and the hook file is absent or tokensave-created.
-/// A user-managed `core.hooksPath` setup is left alone: the user may
-/// deliberately not forward to per-repo hooks.
-fn should_chain_repo_hooks(
-    claiming_hookspath: bool,
-    hooks_dir_is_default: bool,
-    existing_contents: Option<&str>,
-) -> bool {
-    if existing_contents.is_some_and(|c| c.contains(HOOK_MARKER_CHAIN)) {
-        return false;
-    }
-    claiming_hookspath
-        || (hooks_dir_is_default
-            && existing_contents
-                .is_none_or(|c| c.contains(HOOK_MARKER) || c.contains(HOOK_MARKER_CHECKOUT)))
-}
-
-/// The hook snippet appended to (or written as) the post-commit script.
-fn post_commit_snippet(tokensave_bin: &str) -> String {
-    let bin = tokensave_bin.replace('\\', "/");
-    format!(
-        "{HOOK_MARKER}\n\
-         {bin} sync >/dev/null 2>&1 &\n"
-    )
-}
-
-/// The hook snippet appended to (or written as) the post-checkout script.
-///
-/// Runs `tokensave init` in the background on the initial checkout of a fresh
-/// clone — git passes the all-zeros sentinel as the previous HEAD in that case.
-/// On an ordinary **branch** checkout (git passes flag `$3 == 1`) it runs
-/// `tokensave branch add` to transparently track the just-checked-out branch;
-/// that is a no-op when the branch is already tracked or is the default branch.
-/// File checkouts (`$3 == 0`) trigger nothing.
-fn post_checkout_snippet(tokensave_bin: &str) -> String {
-    let bin = tokensave_bin.replace('\\', "/");
-    format!(
-        "{HOOK_MARKER_CHECKOUT}\n\
-         if [ \"$1\" = \"0000000000000000000000000000000000000000\" ]; then\n\
-         \t{bin} init >/dev/null 2>&1 &\n\
-         elif [ \"$3\" = \"1\" ]; then\n\
-         \t{bin} branch add >/dev/null 2>&1 &\n\
-         fi\n"
-    )
-}
-
-/// Append `snippet` to an existing hook file (creating it with a `#!/bin/sh`
-/// shebang first if absent), then mark it executable on Unix. Prints an error
-/// and returns `false` on any I/O failure. Idempotency (skipping when the
-/// tokensave marker is already present) is the caller's responsibility.
-fn write_global_hook(hook_path: &Path, snippet: &str) -> bool {
-    if hook_path.exists() {
-        use std::io::Write;
-        let Ok(mut f) = std::fs::OpenOptions::new().append(true).open(hook_path) else {
-            eprintln!(
-                "  \x1b[31m✘\x1b[0m Failed to open {} for writing",
-                hook_path.display()
-            );
-            return false;
-        };
-        if write!(f, "\n{snippet}").is_err() {
-            eprintln!(
-                "  \x1b[31m✘\x1b[0m Failed to write to {}",
-                hook_path.display()
-            );
-            return false;
-        }
-    } else {
-        let contents = format!("#!/bin/sh\n{snippet}");
-        if std::fs::write(hook_path, contents).is_err() {
-            eprintln!(
-                "  \x1b[31m✘\x1b[0m Failed to create {}",
-                hook_path.display()
-            );
-            return false;
-        }
-    }
-
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        let _ = std::fs::set_permissions(hook_path, std::fs::Permissions::from_mode(0o755));
-    }
-
-    true
-}
-
-/// Action decided by [`decide_hook_action`]: what the caller should do
-/// given the user-supplied mode and the current state of the global
-/// `post-commit` hook file.
-#[derive(Debug, PartialEq, Eq)]
-pub(crate) enum HookAction {
-    /// Hook is already installed (marker present) — nothing to do. The
-    /// caller may still print an informational message.
-    AlreadyInstalled,
-    /// Skip the install entirely (mode `No`, or default-mode non-TTY).
-    Skip,
-    /// Show the interactive prompt and act on the answer.
-    Prompt,
-    /// Install the hook now (no prompt).
-    Install,
-}
-
-/// Pure decision: figure out what to do for the global post-commit hook
-/// given the requested mode and the hook file's current contents (`None`
-/// when the file does not exist or could not be read). The caller
-/// handles all I/O.
-pub(crate) fn decide_hook_action(mode: GitHookMode, hook_contents: Option<&str>) -> HookAction {
-    if hook_contents.is_some_and(|c| c.contains(HOOK_MARKER)) {
-        return HookAction::AlreadyInstalled;
-    }
-
-    match mode {
-        GitHookMode::Default if atty_stdin() => HookAction::Prompt,
-        GitHookMode::Default | GitHookMode::No => HookAction::Skip,
-        GitHookMode::Yes => HookAction::Install,
-    }
-}
-
-/// If a global git `post-commit` hook is not already set up for tokensave,
-/// interactively asks the user whether to install one. Silently succeeds if
-/// the hook is already present, if stdin is not a terminal, or if the user
-/// declines. The `mode` argument lets the caller pre-decide the answer so
-/// scripted installs do not have to drive an interactive prompt.
-pub fn offer_git_post_commit_hook(tokensave_bin: &str, mode: GitHookMode) {
-    let Some(home) = home_dir() else { return };
-
-    // Determine the global hooks directory by reading core.hooksPath from
-    // the global gitconfig file(s). Falls back to ~/.config/git/hooks/.
-    let hooks_dir = read_global_hooks_path(&home);
-
-    let default_hooks_dir = home.join(".config").join("git").join("hooks");
-    let (hooks_dir, need_set_hookspath) = match hooks_dir {
-        Some(dir) => (dir, false),
-        None => (default_hooks_dir.clone(), true),
-    };
-    let hooks_dir_is_default = hooks_dir == default_hooks_dir;
-
-    // Issue #164: a global core.hooksPath makes git ignore every repo's
-    // .git/hooks/, where init.templateDir hooks are copied. tokensave's
-    // hooks chain to the repo's own hooks (below) so nothing stops
-    // running, but if we're about to claim core.hooksPath and the user
-    // relies on a hook template, say so up front.
-    if need_set_hookspath {
-        let template_dir = [
-            home.join(".gitconfig"),
-            home.join(".config").join("git").join("config"),
-        ]
-        .iter()
-        .find_map(|p| parse_gitconfig_value(p, "init", "templatedir"));
-        if let Some(dir) = template_dir {
-            eprintln!(
-                "  \x1b[33m⚠\x1b[0m git init.templateDir is set ({dir}). Installing sets a global \
-                 core.hooksPath, which makes git skip each repository's .git/hooks/. tokensave's \
-                 global hooks forward to the repository's own hooks so they keep running."
-            );
-        }
-    }
-
-    let hook_path = hooks_dir.join("post-commit");
-
-    // Read existing contents once so the decision is pure and the
-    // install path can append without re-reading.
-    let existing_contents: Option<String> = if hook_path.exists() {
-        std::fs::read_to_string(&hook_path).ok()
-    } else {
-        None
-    };
-
-    // Whether to (re)write the post-commit hook. The post-checkout hook is
-    // installed alongside it under the same opt-in, with its own marker, so a
-    // pre-existing post-commit install still gains post-checkout on the next run.
-    let install_post_commit = match decide_hook_action(mode, existing_contents.as_deref()) {
-        HookAction::AlreadyInstalled => {
-            eprintln!("  Global git post-commit hook already contains tokensave, skipping");
-            false
-        }
-        HookAction::Skip => {
-            // Mode `No` (or default-mode non-TTY). Stay quiet — script
-            // callers asked for no output here.
-            return;
-        }
-        HookAction::Prompt => {
-            // TTY + default mode: ask, and bail entirely if the user declines.
-            eprintln!();
-            eprint!(
-                "Install global git \x1b[1mpost-commit\x1b[0m + \x1b[1mpost-checkout\x1b[0m hooks to auto-run \x1b[1mtokensave sync\x1b[0m after each commit and \x1b[1mtokensave init\x1b[0m after a fresh clone? [y/N] "
-            );
-            let mut answer = String::new();
-            if std::io::stdin().read_line(&mut answer).is_err() {
-                return;
-            }
-            if !matches!(answer.trim(), "y" | "Y" | "yes" | "Yes") {
-                eprintln!("  Skipped git hooks");
-                return;
-            }
-            true
-        }
-        HookAction::Install => true,
-    };
-
-    // Create the hooks directory if needed.
-    if let Err(e) = std::fs::create_dir_all(&hooks_dir) {
-        eprintln!(
-            "  \x1b[31m✘\x1b[0m Failed to create {}: {e}",
-            hooks_dir.display()
-        );
-        return;
-    }
-
-    // If no global hooksPath was configured, set it in ~/.gitconfig.
-    if need_set_hookspath {
-        let gitconfig_path = home.join(".gitconfig");
-        if let Err(msg) = set_global_hooks_path(&gitconfig_path, &hooks_dir) {
-            eprintln!("  \x1b[31m✘\x1b[0m {msg} — hook not installed");
-            return;
-        }
-        eprintln!(
-            "\x1b[32m✔\x1b[0m Set git core.hooksPath to {}",
-            hooks_dir.display()
-        );
-    }
-
-    // Issue #164: chain to the repo's own hook before tokensave's snippet
-    // so hooks in .git/hooks/ (e.g. from init.templateDir) keep running.
-    // Also retrofits tokensave-owned hook files from earlier versions.
-    if should_chain_repo_hooks(
-        need_set_hookspath,
-        hooks_dir_is_default,
-        existing_contents.as_deref(),
-    ) {
-        write_global_hook(&hook_path, &chain_repo_hook_snippet("post-commit"));
-    }
-
-    if install_post_commit && write_global_hook(&hook_path, &post_commit_snippet(tokensave_bin)) {
-        eprintln!(
-            "\x1b[32m✔\x1b[0m Installed global git post-commit hook at {}",
-            hook_path.display()
-        );
-    }
-
-    // Install the post-checkout hook so a fresh clone auto-initializes. Its
-    // marker is independent of post-commit's, so this is skipped only when the
-    // post-checkout hook itself is already present.
-    let checkout_path = hooks_dir.join("post-checkout");
-    let checkout_contents = std::fs::read_to_string(&checkout_path).ok();
-    if should_chain_repo_hooks(
-        need_set_hookspath,
-        hooks_dir_is_default,
-        checkout_contents.as_deref(),
-    ) {
-        write_global_hook(&checkout_path, &chain_repo_hook_snippet("post-checkout"));
-    }
-    let checkout_present = checkout_contents.is_some_and(|c| c.contains(HOOK_MARKER_CHECKOUT));
-    if !checkout_present && write_global_hook(&checkout_path, &post_checkout_snippet(tokensave_bin))
-    {
-        eprintln!(
-            "\x1b[32m✔\x1b[0m Installed global git post-checkout hook at {}",
-            checkout_path.display()
-        );
-    }
-
-    // Issue #164 follow-up: claiming a global core.hooksPath disables *every*
-    // hook type in each repo's .git/hooks/, not just the two tokensave owns.
-    // Drop pure forwarders for the remaining client-side hooks so a repo's own
-    // pre-commit / pre-push / commit-msg / … keep running.
-    install_repo_hook_forwarders(&hooks_dir, need_set_hookspath, hooks_dir_is_default);
-}
-
-/// Reads `core.hooksPath` from the global gitconfig files.
-///
-/// Checks `~/.gitconfig` first, then `~/.config/git/config` (the XDG
-/// location). Returns the resolved absolute path, or `None` if the key
-/// is absent from both files.
-fn read_global_hooks_path(home: &Path) -> Option<PathBuf> {
-    let candidates = [
-        home.join(".gitconfig"),
-        home.join(".config").join("git").join("config"),
-    ];
-    for path in &candidates {
-        if let Some(value) = parse_gitconfig_value(path, "core", "hookspath") {
-            let expanded = expand_tilde(&value, home);
-            let p = PathBuf::from(&expanded);
-            if p.is_absolute() {
-                return Some(p);
-            }
-            // Relative paths in gitconfig are relative to the home dir.
-            return Some(home.join(p));
-        }
-    }
-    None
-}
-
-/// Minimal gitconfig parser: finds the value of `key` under `[section]`.
-///
-/// Key matching is case-insensitive (git config keys are case-insensitive).
-/// Handles `key = value`, `key=value`, and quoted values.
-fn parse_gitconfig_value(path: &Path, section: &str, key: &str) -> Option<String> {
-    let contents = std::fs::read_to_string(path).ok()?;
-    let section_lower = section.to_ascii_lowercase();
-    let key_lower = key.to_ascii_lowercase();
-
-    let mut in_section = false;
-    for line in contents.lines() {
-        let trimmed = line.trim();
-        if trimmed.starts_with('[') {
-            // Parse section header: [core], [core "subsection"], etc.
-            let header = trimmed
-                .trim_start_matches('[')
-                .split(']')
-                .next()
-                .unwrap_or("")
-                .trim();
-            let section_name = header.split_whitespace().next().unwrap_or("");
-            in_section = section_name.eq_ignore_ascii_case(&section_lower);
-            continue;
-        }
-        if !in_section {
-            continue;
-        }
-        if trimmed.is_empty() || trimmed.starts_with('#') || trimmed.starts_with(';') {
-            continue;
-        }
-        // Parse key = value
-        if let Some((k, v)) = trimmed.split_once('=') {
-            if k.trim().to_ascii_lowercase() == key_lower {
-                let v = v.trim();
-                // Strip surrounding quotes if present.
-                let v = v
-                    .strip_prefix('"')
-                    .and_then(|s| s.strip_suffix('"'))
-                    .unwrap_or(v);
-                return Some(v.to_string());
-            }
-        }
-    }
-    None
-}
-
-/// Appends `core.hooksPath` to the global gitconfig file, creating it if
-/// necessary. Appends to an existing `[core]` section if one exists,
-/// otherwise adds a new one at the end of the file.
-fn set_global_hooks_path(
-    gitconfig_path: &Path,
-    hooks_dir: &Path,
-) -> std::result::Result<(), String> {
-    let hooks_str = hooks_dir.to_string_lossy().replace('\\', "/");
-    let contents = if gitconfig_path.exists() {
-        std::fs::read_to_string(gitconfig_path)
-            .map_err(|e| format!("Failed to read {}: {e}", gitconfig_path.display()))?
-    } else {
-        String::new()
-    };
-
-    let new_contents = insert_gitconfig_value(&contents, "core", "hooksPath", &hooks_str);
-
-    if let Some(parent) = gitconfig_path.parent() {
-        std::fs::create_dir_all(parent)
-            .map_err(|e| format!("Failed to create {}: {e}", parent.display()))?;
-    }
-    std::fs::write(gitconfig_path, new_contents)
-        .map_err(|e| format!("Failed to write {}: {e}", gitconfig_path.display()))?;
-    Ok(())
-}
-
-/// Inserts `key = value` under `[section]` in gitconfig content.
-/// If the section exists, appends the key after the last line of that section.
-/// Otherwise appends a new section at the end.
-fn insert_gitconfig_value(contents: &str, section: &str, key: &str, value: &str) -> String {
-    let section_lower = section.to_ascii_lowercase();
-    let lines: Vec<&str> = contents.lines().collect();
-    let mut result = Vec::with_capacity(lines.len() + 3);
-    let entry = format!("\t{key} = {value}");
-
-    // Find the target section and the line index just before the next section.
-    let mut section_end: Option<usize> = None;
-    let mut in_section = false;
-    for (i, line) in lines.iter().enumerate() {
-        let trimmed = line.trim();
-        if trimmed.starts_with('[') {
-            if in_section {
-                // We've hit the next section — insert before it.
-                section_end = Some(i);
-                break;
-            }
-            let header = trimmed
-                .trim_start_matches('[')
-                .split(']')
-                .next()
-                .unwrap_or("")
-                .trim();
-            let name = header.split_whitespace().next().unwrap_or("");
-            if name.eq_ignore_ascii_case(&section_lower) {
-                in_section = true;
-            }
-        }
-    }
-    if in_section && section_end.is_none() {
-        // Section runs to end of file.
-        section_end = Some(lines.len());
-    }
-
-    if let Some(insert_at) = section_end {
-        for (i, line) in lines.iter().enumerate() {
-            if i == insert_at {
-                result.push(entry.as_str());
-            }
-            result.push(line);
-        }
-        // If inserting at end-of-file.
-        if insert_at == lines.len() {
-            result.push(&entry);
-        }
-    } else {
-        // Section doesn't exist — append it.
-        for line in &lines {
-            result.push(line);
-        }
-        if !contents.is_empty() && !contents.ends_with('\n') {
-            result.push("");
-        }
-        let section_header = format!("[{section}]");
-        // We need to own these strings for the result.
-        // Re-build as a String directly instead.
-        let mut out = result.join("\n");
-        if !out.is_empty() && !out.ends_with('\n') {
-            out.push('\n');
-        }
-        out.push_str(&section_header);
-        out.push('\n');
-        out.push_str(&entry);
-        out.push('\n');
-        return out;
-    }
-
-    let mut out = result.join("\n");
-    if !out.ends_with('\n') {
-        out.push('\n');
-    }
-    out
-}
-
-/// Expand a leading `~` to the given home directory.
-fn expand_tilde(s: &str, home: &Path) -> String {
-    if let Some(rest) = s.strip_prefix("~/") {
-        return home.join(rest).to_string_lossy().replace('\\', "/");
-    }
-    if s == "~" {
-        return home.to_string_lossy().to_string();
-    }
-    s.to_string()
-}
-
-/// Returns true if stdin is connected to a terminal.
-fn atty_stdin() -> bool {
-    use std::io::IsTerminal;
-    std::io::stdin().is_terminal()
-}
-
-#[cfg(test)]
-#[allow(clippy::unwrap_used)]
-mod git_hook_tests {
-    use super::*;
-    use std::path::Path;
-
-    #[test]
-    fn parse_hookspath_basic() {
-        let config = "[core]\n\thooksPath = /home/user/.git-hooks\n";
-        assert_eq!(
-            parse_gitconfig_value_from_str(config, "core", "hookspath"),
-            Some("/home/user/.git-hooks".to_string())
-        );
-    }
-
-    #[test]
-    fn parse_hookspath_quoted() {
-        let config = "[core]\n\thooksPath = \"/home/user/my hooks\"\n";
-        assert_eq!(
-            parse_gitconfig_value_from_str(config, "core", "hookspath"),
-            Some("/home/user/my hooks".to_string())
-        );
-    }
-
-    #[test]
-    fn parse_hookspath_case_insensitive() {
-        let config = "[Core]\n\tHooksPath = /tmp/hooks\n";
-        assert_eq!(
-            parse_gitconfig_value_from_str(config, "core", "hookspath"),
-            Some("/tmp/hooks".to_string())
-        );
-    }
-
-    #[test]
-    fn parse_hookspath_missing() {
-        let config = "[core]\n\tautocrlf = true\n";
-        assert_eq!(
-            parse_gitconfig_value_from_str(config, "core", "hookspath"),
-            None
-        );
-    }
-
-    #[test]
-    fn parse_hookspath_wrong_section() {
-        let config = "[user]\n\thooksPath = /nope\n[core]\n\tautocrlf = true\n";
-        assert_eq!(
-            parse_gitconfig_value_from_str(config, "core", "hookspath"),
-            None
-        );
-    }
-
-    #[test]
-    fn insert_into_existing_section() {
-        let config = "[user]\n\tname = Test\n[core]\n\tautocrlf = true\n";
-        let result = insert_gitconfig_value(config, "core", "hooksPath", "/tmp/hooks");
-        assert!(result.contains("\thooksPath = /tmp/hooks"));
-        assert!(result.contains("[core]"));
-        assert!(result.contains("autocrlf = true"));
-    }
-
-    #[test]
-    fn insert_new_section() {
-        let config = "[user]\n\tname = Test\n";
-        let result = insert_gitconfig_value(config, "core", "hooksPath", "/tmp/hooks");
-        assert!(result.contains("[core]\n\thooksPath = /tmp/hooks"));
-    }
-
-    #[test]
-    fn insert_into_empty_file() {
-        let result = insert_gitconfig_value("", "core", "hooksPath", "/tmp/hooks");
-        assert!(result.contains("[core]\n\thooksPath = /tmp/hooks"));
-    }
-
-    #[test]
-    fn insert_before_next_section() {
-        let config = "[core]\n\tautocrlf = true\n[user]\n\tname = Test\n";
-        let result = insert_gitconfig_value(config, "core", "hooksPath", "/tmp/hooks");
-        // hooksPath should appear after autocrlf but before [user]
-        let hooks_pos = result.find("hooksPath").unwrap();
-        let user_pos = result.find("[user]").unwrap();
-        let autocrlf_pos = result.find("autocrlf").unwrap();
-        assert!(hooks_pos > autocrlf_pos);
-        assert!(hooks_pos < user_pos);
-    }
-
-    #[test]
-    fn expand_tilde_with_slash() {
-        let home = Path::new("/home/test");
-        assert_eq!(expand_tilde("~/hooks", home), "/home/test/hooks");
-    }
-
-    #[test]
-    fn expand_tilde_bare() {
-        let home = Path::new("/home/test");
-        assert_eq!(expand_tilde("~", home), "/home/test");
-    }
-
-    #[test]
-    fn expand_tilde_no_tilde() {
-        let home = Path::new("/home/test");
-        assert_eq!(expand_tilde("/abs/path", home), "/abs/path");
-    }
-
-    #[test]
-    fn decide_hook_action_yes_installs_when_file_missing() {
-        assert_eq!(
-            decide_hook_action(GitHookMode::Yes, None),
-            HookAction::Install
-        );
-    }
-
-    #[test]
-    fn decide_hook_action_yes_installs_when_file_exists_without_marker() {
-        let contents = "#!/bin/sh\necho hello\n";
-        assert_eq!(
-            decide_hook_action(GitHookMode::Yes, Some(contents)),
-            HookAction::Install
-        );
-    }
-
-    #[test]
-    fn decide_hook_action_yes_reports_already_installed_when_marker_present() {
-        let contents = "#!/bin/sh\n# tokensave: auto-sync\n/usr/bin/tokensave sync\n";
-        assert_eq!(
-            decide_hook_action(GitHookMode::Yes, Some(contents)),
-            HookAction::AlreadyInstalled
-        );
-    }
-
-    #[test]
-    fn decide_hook_action_no_skips_even_when_file_missing() {
-        assert_eq!(decide_hook_action(GitHookMode::No, None), HookAction::Skip);
-    }
-
-    #[test]
-    fn post_checkout_snippet_inits_only_on_fresh_clone() {
-        let s = post_checkout_snippet("/usr/local/bin/tokensave");
-        assert!(
-            s.contains(HOOK_MARKER_CHECKOUT),
-            "must carry its idempotency marker, got: {s}"
-        );
-        assert!(
-            s.contains("/usr/local/bin/tokensave init"),
-            "must run `init` with the resolved binary, got: {s}"
-        );
-        assert!(
-            s.contains("0000000000000000000000000000000000000000"),
-            "must guard on the fresh-clone sentinel so branch switches re-route to branch add, got: {s}"
-        );
-        assert!(
-            s.contains("elif [ \"$3\" = \"1\" ]")
-                && s.contains("/usr/local/bin/tokensave branch add"),
-            "must transparently track the branch on a branch checkout (flag $3==1), got: {s}"
-        );
-    }
-
-    #[test]
-    fn write_global_hook_creates_with_shebang_then_appends() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("post-checkout");
-
-        assert!(write_global_hook(&path, "FIRST\n"));
-        let after_create = std::fs::read_to_string(&path).unwrap();
-        assert!(
-            after_create.starts_with("#!/bin/sh\n"),
-            "new hook file must get a shebang, got: {after_create}"
-        );
-        assert!(after_create.contains("FIRST"));
-
-        assert!(write_global_hook(&path, "SECOND\n"));
-        let after_append = std::fs::read_to_string(&path).unwrap();
-        assert!(
-            after_append.contains("FIRST") && after_append.contains("SECOND"),
-            "second write must append, not clobber, got: {after_append}"
-        );
-
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            let mode = std::fs::metadata(&path).unwrap().permissions().mode();
-            assert_eq!(mode & 0o111, 0o111, "hook must be executable");
-        }
-    }
-
-    #[test]
-    fn bare_name_resolves_through_injected_path() {
-        // A bare `tokensave` that resolves via PATH must survive reinstall
-        // (issue #161).
-        let dir = tempfile::TempDir::new().unwrap();
-        std::fs::write(dir.path().join("tokensave"), "").unwrap();
-        let path_var = dir.path().to_string_lossy().to_string();
-        assert!(command_resolves_to_tokensave_in(
-            "tokensave",
-            Some(&path_var)
-        ));
-        assert!(!command_resolves_to_tokensave_in(
-            "tokensave",
-            Some("/nonexistent")
-        ));
-        assert!(!command_resolves_to_tokensave_in("tokensave", None));
-        // Foreign bare names never match regardless of PATH.
-        assert!(!command_resolves_to_tokensave_in(
-            "othertool",
-            Some(&path_var)
-        ));
-    }
-
-    #[test]
-    fn preserve_mcp_command_replaces_stale_or_foreign_commands() {
-        // Nonexistent absolute path: replace.
-        assert_eq!(
-            preserve_mcp_command_str(Some("/nonexistent/dir/tokensave"), "/new/tokensave"),
-            "/new/tokensave"
-        );
-        // Not a tokensave binary at all: replace.
-        assert_eq!(
-            preserve_mcp_command_str(Some("/bin/sh"), "/new/tokensave"),
-            "/new/tokensave"
-        );
-        // No previous entry: use the new path.
-        assert_eq!(
-            preserve_mcp_command_str(None, "/new/tokensave"),
-            "/new/tokensave"
-        );
-    }
-
-    #[test]
-    fn preserve_mcp_command_reads_string_and_array_shapes() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let abs = dir.path().join("tokensave");
-        std::fs::write(&abs, "").unwrap();
-        let abs = abs.to_string_lossy().to_string();
-
-        let string_shape = serde_json::json!(abs);
-        assert_eq!(preserve_mcp_command(Some(&string_shape), "/new/bin"), abs);
-
-        let array_shape = serde_json::json!([abs, "serve"]);
-        assert_eq!(preserve_mcp_command(Some(&array_shape), "/new/bin"), abs);
-    }
-
-    #[test]
-    fn forwarded_hooks_cover_common_types_but_not_tokensave_owned() {
-        // The two hooks tokensave installs itself carry the chain preamble
-        // plus tokensave's action, so they must NOT be in the pure-forwarder
-        // list (that would double-write / conflict).
-        assert!(!FORWARDED_REPO_HOOKS.contains(&"post-commit"));
-        assert!(!FORWARDED_REPO_HOOKS.contains(&"post-checkout"));
-        // The high-value client-side hooks must be forwarded — these are where
-        // husky / pre-commit / lefthook live.
-        for h in ["pre-commit", "pre-push", "commit-msg", "prepare-commit-msg"] {
-            assert!(
-                FORWARDED_REPO_HOOKS.contains(&h),
-                "{h} must be forwarded or a global hooksPath silently disables it"
-            );
-        }
-        // Server-side hooks are irrelevant to a client `core.hooksPath` and
-        // must not be written.
-        for h in ["pre-receive", "update", "post-receive", "proc-receive"] {
-            assert!(!FORWARDED_REPO_HOOKS.contains(&h));
-        }
-    }
-
-    #[test]
-    fn install_repo_hook_forwarders_writes_when_claiming_and_skips_existing() {
-        let dir = tempfile::tempdir().unwrap();
-        // A hook the user already placed in the dir must be preserved verbatim.
-        let user_pre_commit = dir.path().join("pre-commit");
-        std::fs::write(&user_pre_commit, "#!/bin/sh\n# user's own\n").unwrap();
-
-        install_repo_hook_forwarders(dir.path(), true, true);
-
-        // Existing file untouched.
-        assert_eq!(
-            std::fs::read_to_string(&user_pre_commit).unwrap(),
-            "#!/bin/sh\n# user's own\n",
-            "an existing hook must never be clobbered"
-        );
-        // A forwarder was created for a type that had no file, and it chains
-        // to the repo's own hook of the same name.
-        let created = std::fs::read_to_string(dir.path().join("pre-push")).unwrap();
-        assert!(created.starts_with("#!/bin/sh\n"));
-        assert!(created.contains(HOOK_MARKER_CHAIN));
-        assert!(created.contains("/hooks/pre-push"));
-        assert!(created.contains("git rev-parse --git-dir"));
-    }
-
-    #[test]
-    fn install_repo_hook_forwarders_noop_for_user_managed_hookspath() {
-        // Not claiming, and the dir is not tokensave's default → user-managed
-        // core.hooksPath. tokensave must not write anything into it.
-        let dir = tempfile::tempdir().unwrap();
-        install_repo_hook_forwarders(dir.path(), false, false);
-        assert!(
-            !dir.path().join("pre-commit").exists(),
-            "must leave a user-managed hooksPath directory untouched"
-        );
-    }
-
-    #[test]
-    fn chain_snippet_forwards_to_repo_hook_via_git_dir() {
-        let s = chain_repo_hook_snippet("post-checkout");
-        assert!(s.contains(HOOK_MARKER_CHAIN));
-        // Must use --git-dir, not --git-path hooks: the latter resolves
-        // through core.hooksPath and would re-enter the global hook.
-        assert!(s.contains("git rev-parse --git-dir"));
-        assert!(!s.contains("--git-path"));
-        assert!(s.contains("/hooks/post-checkout"));
-        // Args must be forwarded (post-checkout receives old/new/flag).
-        assert!(s.contains("\"$@\""));
-    }
-
-    #[test]
-    fn should_chain_when_claiming_hookspath() {
-        assert!(should_chain_repo_hooks(true, true, None));
-        assert!(should_chain_repo_hooks(true, false, None));
-    }
-
-    #[test]
-    fn should_chain_retrofits_tokensave_owned_default_dir() {
-        // Existing tokensave-created hook in the default dir gains chaining.
-        assert!(should_chain_repo_hooks(
-            false,
-            true,
-            Some("#!/bin/sh\n# tokensave: auto-sync\ntokensave sync &\n")
-        ));
-        // Absent file in the default dir also chains.
-        assert!(should_chain_repo_hooks(false, true, None));
-    }
-
-    #[test]
-    fn should_not_chain_user_managed_hookspath_or_twice() {
-        // User configured their own core.hooksPath with their own hook.
-        assert!(!should_chain_repo_hooks(
-            false,
-            false,
-            Some("#!/bin/sh\nmy-own-hook\n")
-        ));
-        // Non-tokensave hook file in the default dir is user content too.
-        assert!(!should_chain_repo_hooks(
-            false,
-            true,
-            Some("#!/bin/sh\nmy-own-hook\n")
-        ));
-        // Already chained: never append a second preamble.
-        assert!(!should_chain_repo_hooks(
-            true,
-            true,
-            Some("#!/bin/sh\n# tokensave: chain-repo-hook\n")
-        ));
-    }
-
-    #[test]
-    fn decide_hook_action_no_still_reports_already_installed() {
-        // The user explicitly opted out of changes, but we should still
-        // report that the hook is already in place rather than silently
-        // skipping. Caller prints the message.
-        let contents = "# tokensave: auto-sync\nfoo\n";
-        assert_eq!(
-            decide_hook_action(GitHookMode::No, Some(contents)),
-            HookAction::AlreadyInstalled
-        );
-    }
-
-    #[test]
-    fn decide_hook_action_default_skips_when_file_missing() {
-        // On a non-TTY the default mode silently skips. We cannot
-        // guarantee whether `atty_stdin()` is true or false in a test
-        // process, so assert that the result is one of the two valid
-        // outcomes.
-        let action = decide_hook_action(GitHookMode::Default, None);
-        assert!(matches!(action, HookAction::Skip | HookAction::Prompt));
-    }
-
-    #[test]
-    fn decide_hook_action_default_already_installed_wins_over_tty() {
-        let contents = "# tokensave: auto-sync\nfoo\n";
-        assert_eq!(
-            decide_hook_action(GitHookMode::Default, Some(contents)),
-            HookAction::AlreadyInstalled
-        );
-    }
-
-    /// Helper: parse from a string directly (avoids file I/O in tests).
-    fn parse_gitconfig_value_from_str(contents: &str, section: &str, key: &str) -> Option<String> {
-        let section_lower = section.to_ascii_lowercase();
-        let key_lower = key.to_ascii_lowercase();
-        let mut in_section = false;
-        for line in contents.lines() {
-            let trimmed = line.trim();
-            if trimmed.starts_with('[') {
-                let header = trimmed
-                    .trim_start_matches('[')
-                    .split(']')
-                    .next()
-                    .unwrap_or("")
-                    .trim();
-                let section_name = header.split_whitespace().next().unwrap_or("");
-                in_section = section_name.eq_ignore_ascii_case(&section_lower);
-                continue;
-            }
-            if !in_section {
-                continue;
-            }
-            if trimmed.is_empty() || trimmed.starts_with('#') || trimmed.starts_with(';') {
-                continue;
-            }
-            if let Some((k, v)) = trimmed.split_once('=') {
-                if k.trim().to_ascii_lowercase() == key_lower {
-                    let v = v.trim();
-                    let v = v
-                        .strip_prefix('"')
-                        .and_then(|s| s.strip_suffix('"'))
-                        .unwrap_or(v);
-                    return Some(v.to_string());
-                }
-            }
-        }
-        None
-    }
-}
-
 pub fn tool_names() -> Vec<String> {
     get_tool_definitions()
         .iter()
@@ -2713,941 +394,6 @@ pub fn install_tool_perms(wildcard: bool) -> Vec<String> {
         vec![TOKENSAVE_WILDCARD_PERM.to_string()]
     } else {
         expected_tool_perms()
-    }
-}
-
-#[cfg(test)]
-#[allow(clippy::unwrap_used, clippy::expect_used)]
-mod jsonc_tests {
-    use super::*;
-
-    #[test]
-    fn parse_jsonc_plain_json() {
-        let input = r#"{"key": "value", "num": 42}"#;
-        let v = parse_jsonc(input);
-        assert_eq!(v["key"], "value");
-        assert_eq!(v["num"], 42);
-    }
-
-    #[test]
-    fn parse_jsonc_line_comment() {
-        let input = "{\n  // this is a comment\n  \"key\": \"val\"\n}";
-        let v = parse_jsonc(input);
-        assert_eq!(v["key"], "val");
-    }
-
-    #[test]
-    fn parse_jsonc_block_comment() {
-        let input = "{ /* block comment */ \"key\": \"val\" }";
-        let v = parse_jsonc(input);
-        assert_eq!(v["key"], "val");
-    }
-
-    #[test]
-    fn parse_jsonc_trailing_comma_object() {
-        let input = r#"{"a": 1, "b": 2,}"#;
-        let v = parse_jsonc(input);
-        assert_eq!(v["a"], 1);
-        assert_eq!(v["b"], 2);
-    }
-
-    #[test]
-    fn parse_jsonc_trailing_comma_array() {
-        let input = r#"{"items": [1, 2, 3,]}"#;
-        let v = parse_jsonc(input);
-        assert_eq!(v["items"][2], 3);
-    }
-
-    #[test]
-    fn parse_jsonc_combined() {
-        let input = "{\n  // comment\n  \"x\": /* inline */ 99,\n}";
-        let v = parse_jsonc(input);
-        assert_eq!(v["x"], 99);
-    }
-
-    #[test]
-    fn parse_jsonc_url_in_string_not_stripped() {
-        // A URL containing `//` inside a string must NOT be treated as a comment.
-        let input = r#"{"url": "https://example.com/path"}"#;
-        let v = parse_jsonc(input);
-        assert_eq!(v["url"], "https://example.com/path");
-    }
-
-    #[test]
-    fn parse_jsonc_invalid_falls_back_to_empty() {
-        let input = "not valid json at all !!!";
-        let v = parse_jsonc(input);
-        assert_eq!(v, serde_json::json!({}));
-    }
-
-    #[test]
-    fn parse_jsonc_empty_string() {
-        let v = parse_jsonc("");
-        assert_eq!(v, serde_json::json!({}));
-    }
-
-    #[test]
-    fn parse_jsonc_trailing_comma_with_whitespace() {
-        let input = "{\n  \"a\": 1  ,\n}";
-        let v = parse_jsonc(input);
-        assert_eq!(v["a"], 1);
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Regression tests for safe config backup / load / write
-// ---------------------------------------------------------------------------
-#[cfg(test)]
-#[allow(clippy::unwrap_used, clippy::expect_used)]
-mod safe_config_tests {
-    use super::*;
-    use std::fs;
-
-    /// Create a temp directory that is cleaned up on drop.
-    fn tmpdir() -> tempfile::TempDir {
-        tempfile::tempdir().expect("failed to create temp dir")
-    }
-
-    // ----- backup_config_file -----
-
-    #[test]
-    fn backup_returns_none_when_file_missing() {
-        let dir = tmpdir();
-        let path = dir.path().join("nonexistent.json");
-        let result = backup_config_file(&path).unwrap();
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn backup_creates_bak_with_identical_content() {
-        let dir = tmpdir();
-        let path = dir.path().join("config.json");
-        let original = r#"{"existing": "data", "nested": {"key": 1}}"#;
-        fs::write(&path, original).unwrap();
-
-        let backup = backup_config_file(&path)
-            .unwrap()
-            .expect("should create backup");
-        assert!(backup.exists());
-        assert_eq!(fs::read_to_string(&backup).unwrap(), original);
-        // Original is untouched
-        assert_eq!(fs::read_to_string(&path).unwrap(), original);
-    }
-
-    #[test]
-    fn backup_staging_file_is_cleaned_up() {
-        let dir = tmpdir();
-        let path = dir.path().join("config.json");
-        fs::write(&path, "{}").unwrap();
-
-        backup_config_file(&path).unwrap();
-
-        let staging = dir.path().join("config.json.bak.new");
-        assert!(!staging.exists(), ".bak.new staging file should be removed");
-    }
-
-    // ----- load_json_file_strict -----
-
-    #[test]
-    fn strict_load_returns_empty_for_missing_file() {
-        let dir = tmpdir();
-        let path = dir.path().join("nope.json");
-        let val = load_json_file_strict(&path).unwrap();
-        assert_eq!(val, serde_json::json!({}));
-    }
-
-    #[test]
-    fn strict_load_returns_empty_for_blank_file() {
-        let dir = tmpdir();
-        let path = dir.path().join("empty.json");
-        fs::write(&path, "   \n  ").unwrap();
-        let val = load_json_file_strict(&path).unwrap();
-        assert_eq!(val, serde_json::json!({}));
-    }
-
-    #[test]
-    fn strict_load_parses_valid_json() {
-        let dir = tmpdir();
-        let path = dir.path().join("valid.json");
-        fs::write(&path, r#"{"hello": "world", "n": 42}"#).unwrap();
-        let val = load_json_file_strict(&path).unwrap();
-        assert_eq!(val["hello"], "world");
-        assert_eq!(val["n"], 42);
-    }
-
-    #[test]
-    fn strict_load_errors_on_invalid_json() {
-        let dir = tmpdir();
-        let path = dir.path().join("bad.json");
-        fs::write(&path, "not json {{{").unwrap();
-        let err = load_json_file_strict(&path).unwrap_err();
-        let msg = err.to_string();
-        assert!(msg.contains("cannot parse"), "error: {msg}");
-        assert!(
-            msg.contains("bad.json"),
-            "error should mention filename: {msg}"
-        );
-    }
-
-    #[test]
-    fn strict_load_errors_on_truncated_json() {
-        let dir = tmpdir();
-        let path = dir.path().join("trunc.json");
-        fs::write(&path, r#"{"key": "value", "incomplete"#).unwrap();
-        assert!(load_json_file_strict(&path).is_err());
-    }
-
-    // ----- load_jsonc_file_strict -----
-
-    #[test]
-    fn strict_jsonc_load_returns_empty_for_missing() {
-        let dir = tmpdir();
-        let path = dir.path().join("nope.jsonc");
-        let val = load_jsonc_file_strict(&path).unwrap();
-        assert_eq!(val, serde_json::json!({}));
-    }
-
-    #[test]
-    fn strict_jsonc_load_parses_valid_jsonc() {
-        let dir = tmpdir();
-        let path = dir.path().join("settings.json");
-        fs::write(
-            &path,
-            "{\n  // comment\n  \"key\": \"val\",\n  /* block */ \"n\": 1,\n}",
-        )
-        .unwrap();
-        let val = load_jsonc_file_strict(&path).unwrap();
-        assert_eq!(val["key"], "val");
-        assert_eq!(val["n"], 1);
-    }
-
-    #[test]
-    fn strict_jsonc_load_errors_on_garbage() {
-        let dir = tmpdir();
-        let path = dir.path().join("garbage.json");
-        fs::write(&path, "totally not json or jsonc !!!").unwrap();
-        let err = load_jsonc_file_strict(&path).unwrap_err();
-        assert!(err.to_string().contains("cannot parse"));
-    }
-
-    // ----- safe_write_json_file -----
-
-    #[test]
-    fn safe_write_creates_file_from_scratch() {
-        let dir = tmpdir();
-        let path = dir.path().join("new.json");
-        let value = serde_json::json!({"created": true});
-        safe_write_json_file(&path, &value, None).unwrap();
-
-        let written = fs::read_to_string(&path).unwrap();
-        let parsed: serde_json::Value = serde_json::from_str(&written).unwrap();
-        assert_eq!(parsed["created"], true);
-    }
-
-    #[test]
-    fn safe_write_replaces_existing_file_atomically() {
-        let dir = tmpdir();
-        let path = dir.path().join("existing.json");
-        fs::write(&path, r#"{"old": true}"#).unwrap();
-
-        let value = serde_json::json!({"new": true});
-        safe_write_json_file(&path, &value, None).unwrap();
-
-        let parsed: serde_json::Value =
-            serde_json::from_str(&fs::read_to_string(&path).unwrap()).unwrap();
-        assert_eq!(parsed["new"], true);
-        assert!(parsed.get("old").is_none());
-    }
-
-    #[test]
-    fn safe_write_cleans_up_new_file_on_success() {
-        let dir = tmpdir();
-        let path = dir.path().join("config.json");
-        safe_write_json_file(&path, &serde_json::json!({}), None).unwrap();
-
-        let new_path = dir.path().join("config.json.new");
-        assert!(!new_path.exists(), ".new staging file should be removed");
-    }
-
-    #[test]
-    fn safe_write_creates_parent_dirs() {
-        let dir = tmpdir();
-        let path = dir.path().join("deep").join("nested").join("config.json");
-        safe_write_json_file(&path, &serde_json::json!({"deep": true}), None).unwrap();
-        assert!(path.exists());
-    }
-
-    // ----- symlink handling (dotfiles use case, issue: settings.json
-    //       symlinked into a dotfiles repo was replaced by a plain file) -----
-
-    #[test]
-    #[cfg(unix)]
-    fn safe_write_through_symlink_preserves_link_and_updates_target() {
-        use std::os::unix::fs::symlink;
-
-        let dir = tmpdir();
-        let target = dir.path().join("real_target.json");
-        fs::write(&target, r#"{"old": true}"#).unwrap();
-
-        let link = dir.path().join("settings.json");
-        symlink(&target, &link).unwrap();
-
-        safe_write_json_file(&link, &serde_json::json!({"new": true}), None).unwrap();
-
-        // The symlink itself must still be a symlink pointing at the target.
-        let meta = fs::symlink_metadata(&link).unwrap();
-        assert!(
-            meta.file_type().is_symlink(),
-            "link.json should remain a symlink"
-        );
-        assert_eq!(fs::read_link(&link).unwrap(), target);
-
-        // The real target must contain the new content.
-        let parsed: serde_json::Value =
-            serde_json::from_str(&fs::read_to_string(&target).unwrap()).unwrap();
-        assert_eq!(parsed["new"], true);
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn safe_write_through_symlink_target_in_other_dir() {
-        use std::os::unix::fs::symlink;
-
-        let link_dir = tmpdir();
-        let target_dir = tmpdir();
-        let target = target_dir.path().join("dotfiles_settings.json");
-        fs::write(&target, r#"{"old": true}"#).unwrap();
-
-        let link = link_dir.path().join("settings.json");
-        symlink(&target, &link).unwrap();
-
-        safe_write_json_file(&link, &serde_json::json!({"new": true}), None).unwrap();
-
-        assert!(fs::symlink_metadata(&link)
-            .unwrap()
-            .file_type()
-            .is_symlink());
-        let parsed: serde_json::Value =
-            serde_json::from_str(&fs::read_to_string(&target).unwrap()).unwrap();
-        assert_eq!(parsed["new"], true);
-
-        // No leftover staging file next to the symlink or the target.
-        assert!(!target_dir
-            .path()
-            .join("dotfiles_settings.json.new")
-            .exists());
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn safe_write_through_broken_symlink_creates_target() {
-        use std::os::unix::fs::symlink;
-
-        let dir = tmpdir();
-        let target = dir.path().join("not_yet_created.json");
-        let link = dir.path().join("settings.json");
-        symlink(&target, &link).unwrap(); // target does not exist yet
-
-        safe_write_json_file(&link, &serde_json::json!({"created": true}), None).unwrap();
-
-        assert!(fs::symlink_metadata(&link)
-            .unwrap()
-            .file_type()
-            .is_symlink());
-        let parsed: serde_json::Value =
-            serde_json::from_str(&fs::read_to_string(&target).unwrap()).unwrap();
-        assert_eq!(parsed["created"], true);
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn safe_write_through_multi_hop_dangling_chain_preserves_every_hop() {
-        use std::os::unix::fs::symlink;
-
-        let dir = tmpdir();
-        let final_target = dir.path().join("final_target.json"); // never created
-        let intermediate = dir.path().join("intermediate.json");
-        let config = dir.path().join("config.json");
-
-        symlink(&final_target, &intermediate).unwrap(); // intermediate -> missing final_target
-        symlink(&intermediate, &config).unwrap(); // config -> intermediate
-
-        safe_write_json_file(&config, &serde_json::json!({"created": true}), None).unwrap();
-
-        // Every hop in the chain must survive as a symlink — only the
-        // terminal (previously missing) target becomes a regular file.
-        assert!(
-            fs::symlink_metadata(&config)
-                .unwrap()
-                .file_type()
-                .is_symlink(),
-            "config.json should remain a symlink"
-        );
-        assert_eq!(fs::read_link(&config).unwrap(), intermediate);
-        assert!(
-            fs::symlink_metadata(&intermediate)
-                .unwrap()
-                .file_type()
-                .is_symlink(),
-            "intermediate.json should remain a symlink, not be replaced by a regular file"
-        );
-        assert_eq!(fs::read_link(&intermediate).unwrap(), final_target);
-
-        let parsed: serde_json::Value =
-            serde_json::from_str(&fs::read_to_string(&final_target).unwrap()).unwrap();
-        assert_eq!(parsed["created"], true);
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn safe_write_through_cyclic_symlink_fails_safely_without_touching_links() {
-        use std::os::unix::fs::symlink;
-
-        let dir = tmpdir();
-        let a = dir.path().join("a.json");
-        let b = dir.path().join("b.json");
-        symlink(&b, &a).unwrap(); // a -> b
-        symlink(&a, &b).unwrap(); // b -> a (cycle)
-
-        // Must terminate (cycle-detected) rather than hang, and must refuse
-        // to write rather than fall back to renaming over `a` itself — that
-        // fallback would silently destroy the symlink it's meant to protect.
-        let result = safe_write_json_file(&a, &serde_json::json!({"x": true}), None);
-        assert!(result.is_err(), "a cyclic symlink must be rejected");
-
-        assert!(
-            fs::symlink_metadata(&a).unwrap().file_type().is_symlink(),
-            "a.json must remain untouched after a failed resolution"
-        );
-        assert!(
-            fs::symlink_metadata(&b).unwrap().file_type().is_symlink(),
-            "b.json must remain untouched after a failed resolution"
-        );
-        assert_eq!(fs::read_link(&a).unwrap(), b);
-        assert_eq!(fs::read_link(&b).unwrap(), a);
-    }
-
-    /// Builds a chain of `hops` distinct (non-cyclic) symlinks:
-    /// `hop_0 -> hop_1 -> ... -> hop_{hops-1} -> hop_final_missing` (never
-    /// created). Returns `(entry_path, final_missing_target)`.
-    #[cfg(unix)]
-    fn build_dangling_chain(dir: &Path, hops: usize) -> (PathBuf, PathBuf) {
-        use std::os::unix::fs::symlink;
-
-        let final_target = dir.join("hop_final_missing.json");
-        let mut prev = final_target.clone();
-        for i in (0..hops).rev() {
-            let link = dir.join(format!("hop_{i}.json"));
-            symlink(&prev, &link).unwrap();
-            prev = link;
-        }
-        (prev, final_target)
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn safe_write_through_chain_of_exactly_max_hops_succeeds() {
-        let dir = tmpdir();
-        // A chain of exactly MAX_SYMLINK_HOPS symlinks landing on a terminal
-        // (missing) target must still resolve — the hop budget bounds how
-        // many links are *followed*, not how many are merely inspected, so
-        // the terminal check after the last followed hop must still run.
-        let (entry, final_target) = build_dangling_chain(dir.path(), MAX_SYMLINK_HOPS);
-
-        safe_write_json_file(&entry, &serde_json::json!({"x": true}), None).unwrap();
-
-        assert!(
-            fs::symlink_metadata(&entry)
-                .unwrap()
-                .file_type()
-                .is_symlink(),
-            "entry point must remain a symlink"
-        );
-        let parsed: serde_json::Value =
-            serde_json::from_str(&fs::read_to_string(&final_target).unwrap()).unwrap();
-        assert_eq!(parsed["x"], true);
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn safe_write_through_chain_one_hop_past_max_fails_safely() {
-        let dir = tmpdir();
-        // One hop past the budget must still be rejected, and must not fall
-        // back to writing over the entry point.
-        let (entry, _final_target) = build_dangling_chain(dir.path(), MAX_SYMLINK_HOPS + 1);
-
-        let result = safe_write_json_file(&entry, &serde_json::json!({"x": true}), None);
-        assert!(
-            result.is_err(),
-            "a chain one hop past the budget must be rejected"
-        );
-        assert!(
-            fs::symlink_metadata(&entry)
-                .unwrap()
-                .file_type()
-                .is_symlink(),
-            "entry point must remain untouched after a failed resolution"
-        );
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn safe_write_through_excessively_long_dangling_chain_fails_safely() {
-        let dir = tmpdir();
-        // Well past the limit — also must be rejected rather than falling
-        // back to writing over the entry point.
-        let (entry, _final_target) = build_dangling_chain(dir.path(), 50);
-
-        let result = safe_write_json_file(&entry, &serde_json::json!({"x": true}), None);
-        assert!(
-            result.is_err(),
-            "an overlong dangling chain must be rejected"
-        );
-        assert!(
-            fs::symlink_metadata(&entry)
-                .unwrap()
-                .file_type()
-                .is_symlink(),
-            "entry point must remain untouched after a failed resolution"
-        );
-    }
-
-    // ----- write_json_file (convenience wrapper) -----
-
-    #[test]
-    fn write_json_file_creates_backup_automatically() {
-        let dir = tmpdir();
-        let path = dir.path().join("auto.json");
-        fs::write(&path, r#"{"original": true}"#).unwrap();
-
-        write_json_file(&path, &serde_json::json!({"updated": true})).unwrap();
-
-        // .bak should exist with original content
-        let bak = dir.path().join("auto.json.bak");
-        assert!(bak.exists());
-        let backup_content: serde_json::Value =
-            serde_json::from_str(&fs::read_to_string(&bak).unwrap()).unwrap();
-        assert_eq!(backup_content["original"], true);
-    }
-
-    // ----- THE KEY REGRESSION TEST -----
-    // This is the exact bug the fix addresses: load_json_file silently
-    // returned {} on parse failure, and the install wrote {} + tokensave
-    // back, destroying the user's config.
-
-    #[test]
-    fn invalid_json_is_never_silently_replaced() {
-        let dir = tmpdir();
-        let path = dir.path().join("opencode.json");
-        // Simulate a file that serde_json can't parse (e.g. has trailing commas
-        // that the non-strict loader would silently drop).
-        let corrupted =
-            r#"{"mcp": {"other_server": {"url": "http://example.com"},}, "theme": "dark",}"#;
-        fs::write(&path, corrupted).unwrap();
-
-        // The strict loader must refuse to parse this.
-        let err = load_json_file_strict(&path);
-        assert!(err.is_err(), "strict loader must reject invalid JSON");
-
-        // The original file must be completely untouched.
-        assert_eq!(fs::read_to_string(&path).unwrap(), corrupted);
-
-        // Contrast: the old non-strict loader silently returns {} — this
-        // is the exact behavior that destroyed configs.
-        let old_style = load_json_file(&path);
-        assert_eq!(
-            old_style,
-            serde_json::json!({}),
-            "non-strict loader returns empty"
-        );
-    }
-
-    #[test]
-    fn full_install_cycle_preserves_existing_config() {
-        // Simulate the full install cycle: backup → strict load → mutate → safe write.
-        // Existing keys must be preserved.
-        let dir = tmpdir();
-        let path = dir.path().join("config.json");
-        let original = serde_json::json!({
-            "theme": "dark",
-            "mcp": {
-                "existing_server": {"url": "http://localhost:8080"}
-            },
-            "other_setting": [1, 2, 3]
-        });
-        fs::write(&path, serde_json::to_string_pretty(&original).unwrap()).unwrap();
-
-        // Simulate install
-        let backup = backup_config_file(&path).unwrap();
-        let mut config = load_json_file_strict(&path).unwrap();
-        config["mcp"]["tokensave"] = serde_json::json!({
-            "type": "local",
-            "command": ["tokensave", "serve"]
-        });
-        safe_write_json_file(&path, &config, backup.as_deref()).unwrap();
-
-        // Verify
-        let result: serde_json::Value =
-            serde_json::from_str(&fs::read_to_string(&path).unwrap()).unwrap();
-        // Tokensave was added
-        assert!(result["mcp"]["tokensave"].is_object());
-        // Existing keys survived
-        assert_eq!(result["theme"], "dark");
-        assert_eq!(
-            result["mcp"]["existing_server"]["url"],
-            "http://localhost:8080"
-        );
-        assert_eq!(result["other_setting"], serde_json::json!([1, 2, 3]));
-
-        // Backup exists with original content
-        let bak_content: serde_json::Value =
-            serde_json::from_str(&fs::read_to_string(backup.unwrap()).unwrap()).unwrap();
-        assert!(bak_content.get("tokensave").is_none());
-        assert_eq!(bak_content["theme"], "dark");
-    }
-
-    #[test]
-    fn full_install_cycle_aborts_on_corrupt_file() {
-        // If the existing config is corrupt, the install must fail without
-        // touching the file. This is the core regression test.
-        let dir = tmpdir();
-        let path = dir.path().join("config.json");
-        let corrupt_content = "{ this is not valid json at all }}}";
-        fs::write(&path, corrupt_content).unwrap();
-
-        // Backup succeeds (it just copies bytes)
-        let backup = backup_config_file(&path).unwrap();
-        assert!(backup.is_some());
-
-        // Strict load fails
-        let err = load_json_file_strict(&path);
-        assert!(err.is_err());
-
-        // Original file is byte-for-byte unchanged
-        assert_eq!(fs::read_to_string(&path).unwrap(), corrupt_content);
-        // Backup also has the same content
-        assert_eq!(
-            fs::read_to_string(backup.unwrap()).unwrap(),
-            corrupt_content
-        );
-    }
-
-    #[test]
-    fn safe_write_output_is_valid_json() {
-        // Verify the written file is always parseable JSON (round-trip).
-        let dir = tmpdir();
-        let path = dir.path().join("roundtrip.json");
-        let value = serde_json::json!({
-            "unicode": "héllo wörld 🦀",
-            "nested": {"deep": {"array": [1, null, true, "str"]}},
-            "empty_obj": {},
-            "empty_arr": []
-        });
-
-        safe_write_json_file(&path, &value, None).unwrap();
-
-        let raw = fs::read_to_string(&path).unwrap();
-        let reparsed: serde_json::Value =
-            serde_json::from_str(&raw).expect("written file must be valid JSON");
-        assert_eq!(reparsed, value);
-    }
-
-    // ----- remove_managed_rules_file (issue #256 follow-up: symlink safety) -----
-
-    #[test]
-    #[cfg(unix)]
-    fn remove_managed_rules_file_preserves_symlink_removes_target() {
-        use std::os::unix::fs::symlink;
-
-        let link_dir = tmpdir();
-        let target_dir = tmpdir();
-        let target = target_dir.path().join("dotfiles_tokensave.md");
-        fs::write(&target, "old rules").unwrap();
-
-        let link = link_dir.path().join("tokensave.md");
-        symlink(&target, &link).unwrap();
-
-        remove_managed_rules_file(&link);
-
-        // The symlink itself must survive — a dotfiles setup that symlinks
-        // this file into a repo must not have that symlink silently deleted.
-        let meta = fs::symlink_metadata(&link).unwrap();
-        assert!(meta.file_type().is_symlink(), "symlink must be preserved");
-        assert_eq!(fs::read_link(&link).unwrap(), target);
-
-        // The generated content it pointed at must actually be gone —
-        // otherwise uninstall did not really remove the rules.
-        assert!(!target.exists(), "target content must be removed");
-    }
-
-    #[test]
-    fn remove_managed_rules_file_removes_plain_file_and_prunes_empty_dir() {
-        let dir = tmpdir();
-        let rules_dir = dir.path().join("rules");
-        fs::create_dir_all(&rules_dir).unwrap();
-        let path = rules_dir.join("tokensave.md");
-        fs::write(&path, "rules").unwrap();
-
-        remove_managed_rules_file(&path);
-
-        assert!(!path.exists());
-        assert!(!rules_dir.exists(), "now-empty rules/ dir should be pruned");
-    }
-
-    // ----- remove_legacy_rules_block (issue #256 follow-up: adjacent
-    //       headings, atomic writes, error propagation) -----
-
-    const CLAUDE_MARKER: &str = "## MANDATORY: No Explore Agents When Tokensave Is Available";
-    const CLAUDE_SUBHEADING: &str =
-        "## When you spawn an Explore agent in a tokensave-enabled project";
-
-    #[test]
-    fn remove_legacy_rules_block_noop_when_file_missing() {
-        let dir = tmpdir();
-        let path = dir.path().join("CLAUDE.md");
-        remove_legacy_rules_block(&path, CLAUDE_MARKER, &[CLAUDE_SUBHEADING]).unwrap();
-        assert!(!path.exists());
-    }
-
-    #[test]
-    fn remove_legacy_rules_block_noop_when_marker_absent() {
-        let dir = tmpdir();
-        let path = dir.path().join("CLAUDE.md");
-        fs::write(&path, "# My notes\n\nNothing to do with tokensave here.\n").unwrap();
-
-        remove_legacy_rules_block(&path, CLAUDE_MARKER, &[CLAUDE_SUBHEADING]).unwrap();
-
-        assert_eq!(
-            fs::read_to_string(&path).unwrap(),
-            "# My notes\n\nNothing to do with tokensave here.\n"
-        );
-    }
-
-    #[test]
-    fn remove_legacy_rules_block_preserves_adjacent_user_heading_mentioning_tokensave() {
-        // Regression: the old heuristic decided a sub-heading belonged to
-        // our block by checking `heading_line.contains("tokensave")`, so a
-        // user heading that merely mentions tokensave right after our block
-        // was swallowed as if it were part of it. Matching by the exact
-        // known sub-heading text instead must leave it alone.
-        let dir = tmpdir();
-        let path = dir.path().join("CLAUDE.md");
-        fs::write(
-            &path,
-            format!(
-                "# My notes\n\n\
-                 Some custom content.\n\n\
-                 {CLAUDE_MARKER}\n\n\
-                 legacy body text\n\n\
-                 {CLAUDE_SUBHEADING}\n\n\
-                 legacy sub-body text\n\n\
-                 ## My tokensave workflow notes\n\n\
-                 This is MY content about tokensave, not the installed block.\n"
-            ),
-        )
-        .unwrap();
-
-        remove_legacy_rules_block(&path, CLAUDE_MARKER, &[CLAUDE_SUBHEADING]).unwrap();
-
-        let result = fs::read_to_string(&path).unwrap();
-        assert!(!result.contains(CLAUDE_MARKER));
-        assert!(!result.contains(CLAUDE_SUBHEADING));
-        assert!(result.contains("Some custom content."));
-        assert!(
-            result.contains("## My tokensave workflow notes"),
-            "adjacent user heading must survive: {result}"
-        );
-        assert!(result.contains("This is MY content about tokensave, not the installed block."));
-    }
-
-    #[test]
-    fn remove_legacy_rules_block_leaves_only_custom_content_when_block_is_appended_at_eof() {
-        let dir = tmpdir();
-        let path = dir.path().join("CLAUDE.md");
-        fs::write(
-            &path,
-            format!(
-                "# My notes\n\nSome custom content.\n\n{CLAUDE_MARKER}\n\nlegacy body\n\n{CLAUDE_SUBHEADING}\n\nlegacy sub-body\n"
-            ),
-        )
-        .unwrap();
-
-        remove_legacy_rules_block(&path, CLAUDE_MARKER, &[CLAUDE_SUBHEADING]).unwrap();
-
-        assert_eq!(
-            fs::read_to_string(&path).unwrap(),
-            "# My notes\n\nSome custom content.\n"
-        );
-    }
-
-    #[test]
-    fn remove_legacy_rules_block_removes_file_when_migration_leaves_it_empty() {
-        let dir = tmpdir();
-        let path = dir.path().join("CLAUDE.md");
-        fs::write(
-            &path,
-            format!("{CLAUDE_MARKER}\n\nlegacy body\n\n{CLAUDE_SUBHEADING}\n\nlegacy sub-body\n"),
-        )
-        .unwrap();
-
-        remove_legacy_rules_block(&path, CLAUDE_MARKER, &[CLAUDE_SUBHEADING]).unwrap();
-
-        assert!(!path.exists(), "file with nothing left should be removed");
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn remove_legacy_rules_block_preserves_symlink_when_migration_empties_file() {
-        use std::os::unix::fs::symlink;
-
-        let link_dir = tmpdir();
-        let target_dir = tmpdir();
-        let target = target_dir.path().join("dotfiles_claude_md");
-        fs::write(&target, format!("{CLAUDE_MARKER}\n\nlegacy body\n")).unwrap();
-
-        let link = link_dir.path().join("CLAUDE.md");
-        symlink(&target, &link).unwrap();
-
-        remove_legacy_rules_block(&link, CLAUDE_MARKER, &[CLAUDE_SUBHEADING]).unwrap();
-
-        // The symlink itself must survive a dotfiles-managed setup.
-        let meta = fs::symlink_metadata(&link).unwrap();
-        assert!(meta.file_type().is_symlink(), "symlink must be preserved");
-        assert_eq!(fs::read_link(&link).unwrap(), target);
-
-        // The obsolete legacy content it pointed at must actually be gone.
-        assert!(!target.exists(), "legacy target content must be removed");
-    }
-
-    #[test]
-    fn remove_legacy_rules_block_creates_backup() {
-        let dir = tmpdir();
-        let path = dir.path().join("CLAUDE.md");
-        let original = format!("Custom.\n\n{CLAUDE_MARKER}\n\nlegacy body\n");
-        fs::write(&path, &original).unwrap();
-
-        remove_legacy_rules_block(&path, CLAUDE_MARKER, &[CLAUDE_SUBHEADING]).unwrap();
-
-        let backup = dir.path().join("CLAUDE.md.bak");
-        assert!(backup.exists(), "migration must leave a recoverable backup");
-        assert_eq!(fs::read_to_string(&backup).unwrap(), original);
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn remove_legacy_rules_block_write_failure_returns_err_and_leaves_file_untouched() {
-        // Make the directory read-only so backup_config_file's staging write
-        // (CLAUDE.md.bak.new) fails before safe_write_text_file ever runs,
-        // simulating a permission error / read-only filesystem mid-migration.
-        // The caller (install) must see this as an Err, not a silent success
-        // with stale content left in CLAUDE.md.
-        use std::os::unix::fs::PermissionsExt;
-
-        let dir = tmpdir();
-        let path = dir.path().join("CLAUDE.md");
-        let original = format!("Custom.\n\n{CLAUDE_MARKER}\n\nlegacy body\n");
-        fs::write(&path, &original).unwrap();
-
-        fs::set_permissions(dir.path(), fs::Permissions::from_mode(0o500)).unwrap();
-
-        let result = remove_legacy_rules_block(&path, CLAUDE_MARKER, &[CLAUDE_SUBHEADING]);
-
-        // Restore permissions before any assertion can panic/early-return,
-        // so the tempdir can still be cleaned up on drop.
-        fs::set_permissions(dir.path(), fs::Permissions::from_mode(0o700)).unwrap();
-
-        assert!(result.is_err(), "write failure must surface as Err");
-        assert_eq!(
-            fs::read_to_string(&path).unwrap(),
-            original,
-            "CLAUDE.md must be left untouched when migration fails partway"
-        );
-    }
-
-    #[test]
-    fn safe_write_text_file_rename_failure_leaves_original_untouched() {
-        // Target is a directory, not a file: the `.new` staging write next to
-        // it succeeds (parent dir is writable), but the final rename onto a
-        // directory fails. This isolates the rename-failure branch itself,
-        // uncovered by the migration test above (which only reaches the
-        // earlier backup-write failure).
-        let dir = tmpdir();
-        let path = dir.path().join("target");
-        fs::create_dir(&path).unwrap();
-        fs::write(path.join("existing.txt"), "keep me").unwrap();
-
-        let result = safe_write_text_file(&path, "new content");
-
-        assert!(result.is_err(), "rename onto a directory must fail");
-        assert!(
-            path.is_dir(),
-            "original directory must survive a failed rename"
-        );
-        assert_eq!(
-            fs::read_to_string(path.join("existing.txt")).unwrap(),
-            "keep me"
-        );
-        assert!(
-            !dir.path().join("target.new").exists(),
-            "the failed .new staging file must be cleaned up"
-        );
-    }
-}
-
-#[cfg(test)]
-#[allow(clippy::unwrap_used, clippy::expect_used)]
-mod path_normalize_tests {
-    use super::*;
-
-    #[test]
-    fn normalizes_windows_backslashes() {
-        assert_eq!(
-            normalize_path_separators(r"C:\Users\dev\scoop\shims\tokensave.exe"),
-            "C:/Users/dev/scoop/shims/tokensave.exe"
-        );
-    }
-
-    #[test]
-    fn leaves_unix_paths_unchanged() {
-        assert_eq!(
-            normalize_path_separators("/usr/local/bin/tokensave"),
-            "/usr/local/bin/tokensave"
-        );
-    }
-}
-
-#[cfg(test)]
-#[allow(clippy::unwrap_used)]
-mod install_scope_tests {
-    use super::*;
-    use std::path::PathBuf;
-
-    #[test]
-    fn install_context_base_dir_follows_scope() {
-        let home = PathBuf::from("/home/user");
-        let proj = PathBuf::from("/work/proj");
-
-        let global = InstallContext {
-            home: home.clone(),
-            tokensave_bin: "tokensave".into(),
-            tool_permissions: vec![],
-            scope: InstallScope::Global,
-            force_permission_style: false,
-        };
-        assert_eq!(global.base_dir(), home.as_path());
-        assert!(!global.is_local());
-
-        let local = InstallContext {
-            home: home.clone(),
-            tokensave_bin: "tokensave".into(),
-            tool_permissions: vec![],
-            scope: InstallScope::Local {
-                project_path: proj.clone(),
-            },
-            force_permission_style: false,
-        };
-        assert_eq!(local.base_dir(), proj.as_path());
-        assert!(local.is_local());
     }
 }
